@@ -1,28 +1,36 @@
 {
     "name": "getgrav/grav",
-    "type": "library",
+    "type": "project",
     "description": "Modern, Crazy Fast, Ridiculously Easy and Amazingly Powerful Flat-File CMS",
     "keywords": ["cms","flat-file cms","flat cms","flatfile cms","php"],
     "homepage": "http://getgrav.org",
     "license": "MIT",
     "require": {
-        "php": ">=5.4.0",
-        "twig/twig": "~1.16",
+        "php": ">=5.5.9",
+        "twig/twig": "~1.24",
+        "erusev/parsedown": "dev-master as 1.6.0",
         "erusev/parsedown-extra": "~0.7",
-        "symfony/yaml": "~2.7",
-        "symfony/console": "~2.7",
-        "symfony/event-dispatcher": "~2.7",
-        "symfony/var-dumper": "~2.7",
-        "doctrine/cache": "~1.4",
-        "filp/whoops": "1.2.*@dev",
+        "symfony/yaml": "~2.8",
+        "symfony/console": "~2.8",
+        "symfony/event-dispatcher": "~2.8",
+        "symfony/var-dumper": "~2.8",
+        "symfony/polyfill-iconv": "~1.0",
+        "doctrine/cache": "~1.5",
+        "filp/whoops": "~2.0",
         "monolog/monolog": "~1.0",
         "gregwar/image": "~2.0",
-        "ircmaxell/password-compat": "1.0.*",
         "mrclay/minify": "~2.2",
         "donatj/phpuseragentparser": "~0.3",
         "pimple/pimple": "~3.0",
-        "rockettheme/toolbox": "1.1.*",
-        "maximebf/debugbar": "~1.10"
+        "rockettheme/toolbox": "~1.2",
+        "maximebf/debugbar": "~1.10",
+        "ext-mbstring": "*",
+        "ext-openssl": "*",
+        "ext-curl": "*"
+    },
+    "require-dev": {
+        "codeception/codeception": "^2.1",
+        "fzaninotto/faker": "^1.5"
     },
     "autoload": {
         "psr-4": {
@@ -32,5 +40,15 @@
     },
     "archive": {
         "exclude": ["VERSION"]
+    },
+    "scripts": {
+        "post-create-project-cmd": "bin/grav install",
+        "test": "vendor/bin/codecept run unit",
+        "test-windows": "vendor\\bin\\codecept run unit"
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-develop": "1.x-dev"
+        }
     }
 }
