@@ -8,6 +8,7 @@ use Grav\Common\Config\Config;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use RegexIterator;
+use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 
 define('CSS_ASSET', true);
 define('JS_ASSET', false);
@@ -64,16 +65,19 @@ class Assets
     protected $js_pipeline = false;
 
     // The asset holding arrays
-    protected $collections = array();
-    protected $css = array();
-    protected $js = array();
-    protected $inline_css = array();
-    protected $inline_js = array();
+    protected $collections = [];
+    protected $css = [];
+    protected $js = [];
+    protected $inline_css = [];
+    protected $inline_js = [];
+    protected $imports = [];
 
     // Some configuration variables
     protected $config;
     protected $base_url;
     protected $timestamp = '';
+    protected $assets_dir;
+    protected $assets_url;
 
     // Default values for pipeline settings
     protected $css_minify = true;
@@ -82,10 +86,15 @@ class Assets
     protected $js_minify = true;
 
     // Arrays to hold assets that should NOT be pipelined
-    protected $css_no_pipeline = array();
-    protected $js_no_pipeline = array();
+    protected $css_no_pipeline = [];
+    protected $js_no_pipeline = [];
 
-    public function __construct(array $options = array())
+    /**
+     * Assets constructor.
+     *
+     * @param array $options
+     */
+    public function __construct(array $options = [])
     {
         // Forward config options
         if ($options) {
@@ -117,7 +126,7 @@ class Assets
         }
 
         // Pipeline requires public dir
-        if (($this->js_pipeline || $this->css_pipeline) && !is_dir(ASSETS_DIR)) {
+        if (($this->js_pipeline || $this->css_pipeline) && !is_dir($this->assets_dir)) {
             throw new \Exception('Assets: Public dir not found');
         }
 
@@ -161,7 +170,6 @@ class Assets
             $this->timestamp = '?' . self::getGrav()['cache']->getKey();
         }
 
-
         return $this;
     }
 
@@ -175,11 +183,16 @@ class Assets
         $base_url = self::getGrav()['base_url'];
         $asset_config = (array)$config->get('system.assets');
 
+        /** @var UniformResourceLocator $locator */
+        $locator = self::$grav['locator'];
+        $this->assets_dir = $locator->findResource('asset://') . DS;
+        $this->assets_url = $locator->findResource('asset://', false);
+
         $this->config($asset_config);
         $this->base_url = $base_url . '/';
 
         // Register any preconfigured collections
-        foreach ($config->get('system.assets.collections') as $name => $collection) {
+        foreach ($config->get('system.assets.collections', []) as $name => $collection) {
             $this->registerCollection($name, (array)$collection);
         }
     }
@@ -196,7 +209,7 @@ class Assets
      *
      * @return $this
      */
-    public function add($asset, $priority = null, $pipeline = null)
+    public function add($asset, $priority = null, $pipeline = true)
     {
         // More than one asset
         if (is_array($asset)) {
@@ -230,21 +243,23 @@ class Assets
      * You may add more than one asset passing an array as argument.
      *
      * @param  mixed $asset
-     * @param  int $priority the priority, bigger comes first
-     * @param  bool $pipeline false if this should not be pipelined
-     * @param null $group
+     * @param  int   $priority the priority, bigger comes first
+     * @param  bool  $pipeline false if this should not be pipelined
+     * @param null   $group
      *
      * @return $this
      */
-    public function addCss($asset, $priority = null, $pipeline = null, $group = null)
+    public function addCss($asset, $priority = null, $pipeline = true, $group = null)
     {
         if (is_array($asset)) {
             foreach ($asset as $a) {
                 $this->addCss($a, $priority, $pipeline, $group);
             }
+
             return $this;
         } elseif (isset($this->collections[$asset])) {
-            $this->add($this->collections[$asset], $priority, $pipeline, $group);
+            $this->addCss($this->collections[$asset], $priority, $pipeline, $group);
+
             return $this;
         }
 
@@ -252,17 +267,21 @@ class Assets
             $asset = $this->buildLocalLink($asset);
         }
 
+        // Check for existence
+        if ($asset === false) {
+            return $this;
+        }
+
         $data = [
             'asset'    => $asset,
             'priority' => intval($priority ?: 10),
             'order'    => count($this->css),
-            'pipeline' => $pipeline ?: true,
-            'group' => $group ?: 'head'
+            'pipeline' => (bool)$pipeline,
+            'group'    => $group ?: 'head'
         ];
 
         // check for dynamic array and merge with defaults
-        $count_args = func_num_args();
-        if (func_num_args() == 2) {
+        if (func_num_args() > 1) {
             $dynamic_arg = func_get_arg(1);
             if (is_array($dynamic_arg)) {
                 $data = array_merge($data, $dynamic_arg);
@@ -283,22 +302,25 @@ class Assets
      * It checks for duplicates.
      * You may add more than one asset passing an array as argument.
      *
-     * @param  mixed $asset
-     * @param  int $priority the priority, bigger comes first
-     * @param  bool $pipeline false if this should not be pipelined
-     * @param  string $loading how the asset is loaded (async/defer)
-     * @param  string $group name of the group
+     * @param  mixed  $asset
+     * @param  int    $priority the priority, bigger comes first
+     * @param  bool   $pipeline false if this should not be pipelined
+     * @param  string $loading  how the asset is loaded (async/defer)
+     * @param  string $group    name of the group
+     *
      * @return $this
      */
-    public function addJs($asset, $priority = null, $pipeline = null, $loading = null, $group = null)
+    public function addJs($asset, $priority = null, $pipeline = true, $loading = null, $group = null)
     {
         if (is_array($asset)) {
             foreach ($asset as $a) {
                 $this->addJs($a, $priority, $pipeline, $loading, $group);
             }
+
             return $this;
         } elseif (isset($this->collections[$asset])) {
-            $this->add($this->collections[$asset], $priority, $pipeline, $loading, $group);
+            $this->addJs($this->collections[$asset], $priority, $pipeline, $loading, $group);
+
             return $this;
         }
 
@@ -306,21 +328,25 @@ class Assets
             $asset = $this->buildLocalLink($asset);
         }
 
+        // Check for existence
+        if ($asset === false) {
+            return $this;
+        }
+
         $data = [
             'asset'    => $asset,
             'priority' => intval($priority ?: 10),
             'order'    => count($this->js),
-            'pipeline' => $pipeline ?: true,
+            'pipeline' => (bool)$pipeline,
             'loading'  => $loading ?: '',
-            'group' => $group ?: 'head'
+            'group'    => $group ?: 'head'
         ];
 
         // check for dynamic array and merge with defaults
-        $count_args = func_num_args();
-        if (func_num_args() == 2) {
+        if (func_num_args() > 1) {
             $dynamic_arg = func_get_arg(1);
             if (is_array($dynamic_arg)) {
-               $data = array_merge($data, $dynamic_arg);
+                $data = array_merge($data, $dynamic_arg);
             }
         }
 
@@ -335,16 +361,16 @@ class Assets
     /**
      * Convenience wrapper for async loading of JavaScript
      *
-     * @param      $asset
-     * @param int  $priority
-     * @param bool $pipeline
+     * @param        $asset
+     * @param int    $priority
+     * @param bool   $pipeline
      * @param string $group name of the group
      *
      * @deprecated Please use dynamic method with ['loading' => 'async']
      *
      * @return \Grav\Common\Assets
      */
-    public function addAsyncJs($asset, $priority = null, $pipeline = null, $group = null)
+    public function addAsyncJs($asset, $priority = null, $pipeline = true, $group = null)
     {
         return $this->addJs($asset, $priority, $pipeline, 'async', $group);
     }
@@ -352,16 +378,16 @@ class Assets
     /**
      * Convenience wrapper for deferred loading of JavaScript
      *
-     * @param      $asset
-     * @param int  $priority
-     * @param bool $pipeline
+     * @param        $asset
+     * @param int    $priority
+     * @param bool   $pipeline
      * @param string $group name of the group
      *
      * @deprecated Please use dynamic method with ['loading' => 'defer']
      *
      * @return \Grav\Common\Assets
      */
-    public function addDeferJs($asset, $priority = null, $pipeline = null, $group = null)
+    public function addDeferJs($asset, $priority = null, $pipeline = true, $group = null)
     {
         return $this->addJs($asset, $priority, $pipeline, 'defer', $group);
     }
@@ -373,8 +399,8 @@ class Assets
      * For adding chunks of string-based inline CSS
      *
      * @param  mixed $asset
-     * @param  int $priority the priority, bigger comes first
-     * @param null $group
+     * @param  int   $priority the priority, bigger comes first
+     * @param null   $group
      *
      * @return $this
      */
@@ -383,21 +409,20 @@ class Assets
         $asset = trim($asset);
 
         if (is_a($asset, 'Twig_Markup')) {
-            preg_match(self::HTML_TAG_REGEX, $asset, $matches );
-            if (isset($matches[3]))  {
+            preg_match(self::HTML_TAG_REGEX, $asset, $matches);
+            if (isset($matches[3])) {
                 $asset = $matches[3];
             }
         }
 
         $data = [
-            'priority'  => intval($priority ?: 10),
-            'order'     => count($this->inline_css),
-            'asset'     => $asset,
-            'group'     => $group ?: 'head'
+            'priority' => intval($priority ?: 10),
+            'order'    => count($this->inline_css),
+            'asset'    => $asset,
+            'group'    => $group ?: 'head'
         ];
 
         // check for dynamic array and merge with defaults
-        $count_args = func_num_args();
         if (func_num_args() == 2) {
             $dynamic_arg = func_get_arg(1);
             if (is_array($dynamic_arg)) {
@@ -406,7 +431,7 @@ class Assets
         }
 
         $key = md5($asset);
-        if (is_string($asset) && !array_key_exists($key, $this->inline_css)) {
+        if ($asset && is_string($asset) && !array_key_exists($key, $this->inline_css)) {
             $this->inline_css[$key] = $data;
         }
 
@@ -421,7 +446,7 @@ class Assets
      *
      * @param  mixed $asset
      * @param  int   $priority the priority, bigger comes first
-     * @param string $group name of the group
+     * @param string $group    name of the group
      *
      * @return $this
      */
@@ -430,8 +455,8 @@ class Assets
         $asset = trim($asset);
 
         if (is_a($asset, 'Twig_Markup')) {
-            preg_match(self::HTML_TAG_REGEX, $asset, $matches );
-            if (isset($matches[3]))  {
+            preg_match(self::HTML_TAG_REGEX, $asset, $matches);
+            if (isset($matches[3])) {
                 $asset = $matches[3];
             }
         }
@@ -440,11 +465,10 @@ class Assets
             'asset'    => $asset,
             'priority' => intval($priority ?: 10),
             'order'    => count($this->js),
-            'group' => $group ?: 'head'
+            'group'    => $group ?: 'head'
         ];
 
         // check for dynamic array and merge with defaults
-        $count_args = func_num_args();
         if (func_num_args() == 2) {
             $dynamic_arg = func_get_arg(1);
             if (is_array($dynamic_arg)) {
@@ -453,7 +477,7 @@ class Assets
         }
 
         $key = md5($asset);
-        if (is_string($asset) && !array_key_exists($key, $this->inline_js)) {
+        if ($asset && is_string($asset) && !array_key_exists($key, $this->inline_js)) {
             $this->inline_js[$key] = $data;
         }
 
@@ -464,29 +488,31 @@ class Assets
      * Build the CSS link tags.
      *
      * @param  string $group name of the group
-     * @param  array $attributes
+     * @param  array  $attributes
      *
      * @return string
      */
     public function css($group = 'head', $attributes = [])
     {
-        if (!$this->css) {
+        if (!$this->css && !$this->inline_css) {
             return null;
         }
 
         // Sort array by priorities (larger priority first)
         if (self::getGrav()) {
-            usort($this->css, function ($a, $b) {
+            uasort($this->css, function ($a, $b) {
                 if ($a['priority'] == $b['priority']) {
                     return $b['order'] - $a['order'];
                 }
+
                 return $a['priority'] - $b['priority'];
             });
 
-            usort($this->inline_css, function ($a, $b) {
+            uasort($this->inline_css, function ($a, $b) {
                 if ($a['priority'] == $b['priority']) {
                     return $b['order'] - $a['order'];
                 }
+
                 return $a['priority'] - $b['priority'];
             });
         }
@@ -537,28 +563,30 @@ class Assets
      * Build the JavaScript script tags.
      *
      * @param  string $group name of the group
-     * @param  array $attributes
+     * @param  array  $attributes
      *
      * @return string
      */
     public function js($group = 'head', $attributes = [])
     {
-        if (!$this->js) {
+        if (!$this->js && !$this->inline_js) {
             return null;
         }
 
         // Sort array by priorities (larger priority first)
-        usort($this->js, function ($a, $b) {
+        uasort($this->js, function ($a, $b) {
             if ($a['priority'] == $b['priority']) {
                 return $b['order'] - $a['order'];
             }
+
             return $a['priority'] - $b['priority'];
         });
 
-        usort($this->inline_js, function ($a, $b) {
+        uasort($this->inline_js, function ($a, $b) {
             if ($a['priority'] == $b['priority']) {
                 return $b['order'] - $a['order'];
             }
+
             return $a['priority'] - $b['priority'];
         });
 
@@ -577,7 +605,7 @@ class Assets
             }
             foreach ($this->js_no_pipeline as $file) {
                 if ($group && $file['group'] == $group) {
-                    $output .= '<script src="' . $file['asset'] . $this->timestamp . '"' . $attributes . ' ' . $file['loading']. '></script>' . "\n";
+                    $output .= '<script src="' . $file['asset'] . $this->timestamp . '"' . $attributes . ' ' . $file['loading'] . '></script>' . "\n";
                 }
             }
         } else {
@@ -603,9 +631,11 @@ class Assets
     }
 
     /**
-     * Minify and concatenate CSS.
+     * Minify and concatenate CSS
      *
-     * @return string
+     * @param string $group
+     *
+     * @return bool|string
      */
     protected function pipelineCss($group = 'head')
     {
@@ -621,8 +651,8 @@ class Assets
 
         $file = md5(json_encode($this->css) . $this->css_minify . $this->css_rewrite . $group) . '.css';
 
-        $relative_path = "{$this->base_url}" . basename(ASSETS_DIR) . "/{$file}";
-        $absolute_path = ASSETS_DIR . $file;
+        $relative_path = "{$this->base_url}{$this->assets_url}/{$file}";
+        $absolute_path = $this->assets_dir . $file;
 
         // If pipeline exist return it
         if (file_exists($absolute_path)) {
@@ -664,6 +694,7 @@ class Assets
         // Write file
         if (strlen(trim($buffer)) > 0) {
             file_put_contents($absolute_path, $buffer);
+
             return $relative_path . $key;
         } else {
             return false;
@@ -672,6 +703,8 @@ class Assets
 
     /**
      * Minify and concatenate JS files.
+     *
+     * @param string $group
      *
      * @return string
      */
@@ -689,8 +722,8 @@ class Assets
 
         $file = md5(json_encode($this->js) . $this->js_minify . $group) . '.js';
 
-        $relative_path = "{$this->base_url}" . basename(ASSETS_DIR) . "/{$file}";
-        $absolute_path = ASSETS_DIR . $file;
+        $relative_path = "{$this->base_url}{$this->assets_url}/{$file}";
+        $absolute_path = $this->assets_dir . $file;
 
         // If pipeline exist return it
         if (file_exists($absolute_path)) {
@@ -722,6 +755,7 @@ class Assets
         // Write file
         if (strlen(trim($buffer)) > 0) {
             file_put_contents($absolute_path, $buffer);
+
             return $relative_path . $key;
         } else {
             return false;
@@ -767,9 +801,7 @@ class Assets
      */
     public function exists($asset)
     {
-        if (isset($this->collections[$asset]) ||
-            isset($this->css[$asset]) ||
-            isset($this->js[$asset])) {
+        if (isset($this->collections[$asset]) || isset($this->css[$asset]) || isset($this->js[$asset])) {
             return true;
         } else {
             return false;
@@ -811,7 +843,8 @@ class Assets
      */
     public function resetJs()
     {
-        $this->js = array();
+        $this->js = [];
+        $this->inline_js = [];
 
         return $this;
     }
@@ -823,15 +856,28 @@ class Assets
      */
     public function resetCss()
     {
-        $this->css = array();
+        $this->css = [];
+        $this->inline_css = [];
 
         return $this;
     }
 
     /**
-     * Add all CSS assets within $directory (relative to public dir).
+     * Add all JavaScript assets within $directory
      *
-     * @param  string $directory Relative to $this->public_dir
+     * @param  string $directory Relative to the Grav root path, or a stream identifier
+     *
+     * @return $this
+     */
+    public function addDirJs($directory)
+    {
+        return $this->addDir($directory, self::JS_REGEX);
+    }
+
+    /**
+     * Add all CSS assets within $directory
+     *
+     * @param  string $directory Relative to the Grav root path, or a stream identifier
      *
      * @return $this
      */
@@ -843,7 +889,7 @@ class Assets
     /**
      * Add all assets matching $pattern within $directory.
      *
-     * @param  string $directory Relative to $this->public_dir
+     * @param  string $directory Relative to the Grav root path, or a stream identifier
      * @param  string $pattern   (regex)
      *
      * @return $this
@@ -851,13 +897,15 @@ class Assets
      */
     public function addDir($directory, $pattern = self::DEFAULT_REGEX)
     {
-        // Check if public_dir exists
-        if (!is_dir(ASSETS_DIR)) {
-            throw new Exception('Assets: Public dir not found');
+        $root_dir = rtrim(ROOT_DIR, '/');
+
+        // Check if $directory is a stream.
+        if (strpos($directory, '://')) {
+            $directory = self::$grav['locator']->findResource($directory, null);
         }
 
         // Get files
-        $files = $this->rglob(ASSETS_DIR . DIRECTORY_SEPARATOR . $directory, $pattern, ASSETS_DIR);
+        $files = $this->rglob($root_dir . DIRECTORY_SEPARATOR . $directory, $pattern, $root_dir . '/');
 
         // No luck? Nothing to do
         if (!$files) {
@@ -866,27 +914,25 @@ class Assets
 
         // Add CSS files
         if ($pattern === self::CSS_REGEX) {
-            $this->css = array_unique(array_merge($this->css, $files));
+            foreach ($files as $file) {
+                $this->addCss($file);
+            }
+
             return $this;
         }
 
         // Add JavaScript files
         if ($pattern === self::JS_REGEX) {
-            $this->js = array_unique(array_merge($this->js, $files));
+            foreach ($files as $file) {
+                $this->addJs($file);
+            }
+
             return $this;
         }
 
-        // Unknown pattern. We must poll to know the extension :(
+        // Unknown pattern.
         foreach ($files as $asset) {
-            $info = pathinfo($asset);
-            if (isset($info['extension'])) {
-                $ext = strtolower($info['extension']);
-                if ($ext === 'css' && !in_array($asset, $this->css)) {
-                    $this->css[] = $asset;
-                } elseif ($ext === 'js' && !in_array($asset, $this->js)) {
-                    $this->js[] = $asset;
-                }
-            }
+            $this->add($asset);
         }
 
         return $this;
@@ -903,8 +949,8 @@ class Assets
      */
     protected function isRemoteLink($link)
     {
-        return ('http://' === substr($link, 0, 7) || 'https://' === substr($link, 0, 8)
-            || '//' === substr($link, 0, 2));
+        return ('http://' === substr($link, 0, 7) || 'https://' === substr($link, 0, 8) || '//' === substr($link, 0,
+                2));
     }
 
     /**
@@ -957,18 +1003,18 @@ class Assets
      * Download and concatenate the content of several links.
      *
      * @param  array $links
-     * @param  bool $css
+     * @param  bool  $css
      *
      * @return string
      */
     protected function gatherLinks(array $links, $css = true)
     {
-
-
         $buffer = '';
         $local = true;
 
         foreach ($links as $asset) {
+            $relative_dir = '';
+
             $link = $asset['asset'];
             $relative_path = $link;
 
@@ -1019,8 +1065,8 @@ class Assets
     /**
      * Finds relative CSS urls() and rewrites the URL with an absolute one
      *
-     * @param $file                 the css source file
-     * @param $relative_path        relative path to the css file
+     * @param string $file          the css source file
+     * @param string $relative_path relative path to the css file
      *
      * @return mixed
      */
@@ -1031,23 +1077,19 @@ class Assets
 
         // Find any css url() elements, grab the URLs and calculate an absolute path
         // Then replace the old url with the new one
-        $file = preg_replace_callback(
-            self::CSS_URL_REGEX,
-            function ($matches) use ($relative_path) {
+        $file = preg_replace_callback(self::CSS_URL_REGEX, function ($matches) use ($relative_path) {
 
-                $old_url = $matches[1];
+            $old_url = $matches[1];
 
-                // ensure this is not a data url
-                if (strpos($old_url, 'data:') === 0) {
-                    return $matches[0];
-                }
+            // ensure this is not a data url
+            if (strpos($old_url, 'data:') === 0) {
+                return $matches[0];
+            }
 
-                $new_url = $this->base_url . ltrim(Utils::normalizePath($relative_path . '/' . $old_url), '/');
+            $new_url = $this->base_url . ltrim(Utils::normalizePath($relative_path . '/' . $old_url), '/');
 
-                return str_replace($old_url, $new_url, $matches[0]);
-            },
-            $file
-        );
+            return str_replace($old_url, $new_url, $matches[0]);
+        }, $file);
 
         return $file;
     }
@@ -1061,16 +1103,13 @@ class Assets
      */
     protected function moveImports($file)
     {
-        $this->imports = array();
+        $this->imports = [];
 
-        $file = preg_replace_callback(
-            self::CSS_IMPORT_REGEX,
-            function ($matches) {
-                $this->imports[] = $matches[0];
-                return '';
-            },
-            $file
-        );
+        $file = preg_replace_callback(self::CSS_IMPORT_REGEX, function ($matches) {
+            $this->imports[] = $matches[0];
+
+            return '';
+        }, $file);
 
         return implode("\n", $this->imports) . "\n\n" . $file;
     }
@@ -1086,17 +1125,10 @@ class Assets
      */
     protected function rglob($directory, $pattern, $ltrim = null)
     {
-        $iterator = new RegexIterator(
-            new RecursiveIteratorIterator(
-                new RecursiveDirectoryIterator(
-                    $directory,
-                    FilesystemIterator::SKIP_DOTS
-                )
-            ),
-            $pattern
-        );
+        $iterator = new RegexIterator(new RecursiveIteratorIterator(new RecursiveDirectoryIterator($directory,
+            FilesystemIterator::SKIP_DOTS)), $pattern);
         $offset = strlen($ltrim);
-        $files = array();
+        $files = [];
 
         foreach ($iterator as $file) {
             $files[] = substr($file->getPathname(), $offset);
@@ -1106,17 +1138,38 @@ class Assets
     }
 
     /**
-     * Add all JavaScript assets within $directory.
+     * Sets the state of CSS Pipeline
      *
-     * @param  string $directory Relative to $this->public_dir
-     *
-     * @return $this
+     * @param boolean $value
      */
-    public function addDirJs($directory)
+    public function setCssPipeline($value)
     {
-        return $this->addDir($directory, self::JS_REGEX);
+        $this->css_pipeline = (bool)$value;
     }
 
+    /**
+     * Sets the state of JS Pipeline
+     *
+     * @param boolean $value
+     */
+    public function setJsPipeline($value)
+    {
+        $this->js_pipeline = (bool)$value;
+    }
+
+    /**
+     * Explicitly set's a timestamp for assets
+     *
+     * @param $value
+     */
+    public function setTimestamp($value)
+    {
+        $this->timestamp = '?' . $value;
+    }
+
+    /**
+     * @return string
+     */
     public function __toString()
     {
         return '';
