@@ -26,6 +26,7 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
      *
      * @param  string $key
      * @param  mixed  $args
+     *
      * @return mixed
      */
     public function __call($key, $args)
@@ -79,11 +80,13 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
      * Return nth item.
      *
      * @param int $key
+     *
      * @return mixed|bool
      */
     public function nth($key)
     {
         $items = array_keys($this->items);
+
         return (isset($items[$key])) ? $this->offsetGet($items[$key]) : false;
     }
 
@@ -95,6 +98,7 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
     public function first()
     {
         $items = array_keys($this->items);
+
         return $this->offsetGet(array_shift($items));
     }
 
@@ -106,6 +110,7 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
     public function last()
     {
         $items = array_keys($this->items);
+
         return $this->offsetGet(array_pop($items));
     }
 
@@ -117,11 +122,13 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
     public function reverse()
     {
         $this->items = array_reverse($this->items);
+
         return $this;
     }
 
     /**
      * @param mixed $needle Searched value.
+     *
      * @return string|bool  Key if found, otherwise false.
      */
     public function indexOf($needle)
@@ -131,6 +138,7 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
                 return $key;
             }
         }
+
         return false;
     }
 
@@ -144,7 +152,7 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
         $keys = array_keys($this->items);
         shuffle($keys);
 
-        $new = array();
+        $new = [];
         foreach ($keys as $key) {
             $new[$key] = $this->items[$key];
         }
@@ -159,6 +167,7 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
      *
      * @param int $offset
      * @param int $length
+     *
      * @return $this
      */
     public function slice($offset, $length = null)
@@ -171,12 +180,13 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
     /**
      * Pick one or more random entries.
      *
-     * @param int $num  Specifies how many entries should be picked.
+     * @param int $num Specifies how many entries should be picked.
+     *
      * @return $this
      */
     public function random($num = 1)
     {
-        $this->items = array_intersect_key($this->items, array_flip((array) array_rand($this->items, $num)));
+        $this->items = array_intersect_key($this->items, array_flip((array)array_rand($this->items, $num)));
 
         return $this;
     }
@@ -184,7 +194,8 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
     /**
      * Append new elements to the list.
      *
-     * @param array|Iterator $items  Items to be appended. Existing keys will be overridden with the new values.
+     * @param array|Iterator $items Items to be appended. Existing keys will be overridden with the new values.
+     *
      * @return $this
      */
     public function append($items)
@@ -192,14 +203,17 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
         if ($items instanceof static) {
             $items = $items->toArray();
         }
-        $this->items = array_merge($this->items, (array) $items);
+        $this->items = array_merge($this->items, (array)$items);
 
         return $this;
     }
 
     /**
      * Filter elements from the list
-     * @param  callable|null $callback A function the receives ($value, $key) and must return a boolean to indicate filter status
+     *
+     * @param  callable|null $callback A function the receives ($value, $key) and must return a boolean to indicate
+     *                                 filter status
+     *
      * @return $this
      */
     public function filter(callable $callback = null)
@@ -207,12 +221,36 @@ class Iterator implements \ArrayAccess, \Iterator, \Countable, \Serializable
         foreach ($this->items as $key => $value) {
             if (
                 ($callback && !call_user_func($callback, $value, $key)) ||
-                (!$callback && !(bool) $value)
+                (!$callback && !(bool)$value)
             ) {
                 unset($this->items[$key]);
             }
         }
 
         return $this;
+    }
+
+
+    /**
+     * Sorts elements from the list and returns a copy of the list in the proper order
+     *
+     * @param callable|null $callback
+     *
+     * @param bool          $desc
+     *
+     * @return $this|array
+     * @internal param bool $asc
+     *
+     */
+    public function sort(callable $callback = null, $desc = false)
+    {
+        if (!$callback || !is_callable($callback)) {
+            return $this;
+        }
+
+        $items = $this->items;
+        uasort($items, $callback);
+
+        return !$desc ? $items : array_reverse($items, true);
     }
 }
