@@ -1,12 +1,11 @@
 <?php
 namespace Grav\Common\Markdown;
 
-use Grav\Common\Config\Config;
-use Grav\Common\Debugger;
 use Grav\Common\GravTrait;
-use Grav\Common\Page\Medium\Medium;
+use Grav\Common\Page\Page;
+use Grav\Common\Page\Pages;
 use Grav\Common\Uri;
-use Grav\Common\Utils;
+use RocketTheme\Toolbox\Event\Event;
 
 /**
  * A trait to add some custom processing to the identifyLink() method in Parsedown and ParsedownExtra
@@ -14,13 +13,23 @@ use Grav\Common\Utils;
 trait ParsedownGravTrait
 {
     use GravTrait;
+
+    /** @var Page $page */
     protected $page;
+
+    /** @var Pages $pages */
     protected $pages;
-    protected $base_url;
+
+    /** @var  Uri $uri */
+    protected $uri;
+
     protected $pages_dir;
     protected $special_chars;
-
     protected $twig_link_regex = '/\!*\[(?:.*)\]\((\{([\{%#])\s*(.*?)\s*(?:\2|\})\})\)/';
+    protected $special_protocols = ['xmpp', 'mailto', 'tel', 'sms'];
+
+    public $completable_blocks = [];
+    public $continuable_blocks = [];
 
     /**
      * Initialization function to setup key variables needed by the MarkdownGravLinkTrait
@@ -30,12 +39,14 @@ trait ParsedownGravTrait
      */
     protected function init($page, $defaults)
     {
+        $grav = self::getGrav();
+
         $this->page = $page;
-        $this->pages = self::getGrav()['pages'];
+        $this->pages = $grav['pages'];
+        $this->uri = $grav['uri'];
         $this->BlockTypes['{'] [] = "TwigTag";
-        $this->base_url = rtrim(self::getGrav()['base_url'] . self::getGrav()['pages']->base(), '/');
         $this->pages_dir = self::getGrav()['locator']->findResource('page://');
-        $this->special_chars = array('>' => 'gt', '<' => 'lt', '"' => 'quot');
+        $this->special_chars = ['>' => 'gt', '<' => 'lt', '"' => 'quot'];
 
         if ($defaults === null) {
             $defaults = self::getGrav()['config']->get('system.pages.markdown');
@@ -45,12 +56,76 @@ trait ParsedownGravTrait
         $this->setUrlsLinked($defaults['auto_url_links']);
         $this->setMarkupEscaped($defaults['escape_markup']);
         $this->setSpecialChars($defaults['special_chars']);
+
+        $grav->fireEvent('onMarkdownInitialized', new Event(['markdown' => $this]));
+
     }
+
+    /**
+     * Be able to define a new Block type or override an existing one
+     *
+     * @param $type
+     * @param $tag
+     */
+    public function addBlockType($type, $tag, $continuable = false, $completable = false)
+    {
+        $this->BlockTypes[$type] [] = $tag;
+
+        if ($continuable) {
+            $this->continuable_blocks[] = $tag;
+        }
+
+        if ($completable) {
+            $this->completable_blocks[] = $tag;
+        }
+    }
+
+    /**
+     * Be able to define a new Inline type or override an existing one
+     *
+     * @param $type
+     * @param $tag
+     */
+    public function addInlineType($type, $tag)
+    {
+        $this->InlineTypes[$type] [] = $tag;
+        $this->inlineMarkerList .= $type;
+    }
+
+    /**
+     * Overrides the default behavior to allow for plugin-provided blocks to be continuable
+     *
+     * @param $Type
+     *
+     * @return bool
+     */
+    protected function isBlockContinuable($Type)
+    {
+        $continuable = in_array($Type, $this->continuable_blocks) || method_exists($this, 'block' . $Type . 'Continue');
+
+        return $continuable;
+    }
+
+    /**
+     *  Overrides the default behavior to allow for plugin-provided blocks to be completable
+     *
+     * @param $Type
+     *
+     * @return bool
+     */
+    protected function isBlockCompletable($Type)
+    {
+        $completable = in_array($Type, $this->completable_blocks) || method_exists($this, 'block' . $Type . 'Complete');
+
+        return $completable;
+    }
+
 
     /**
      * Make the element function publicly accessible, Medium uses this to render from Twig
      *
-     * @param  array  $Element
+     * @param  array $Element
+     *
      * @return string markup
      */
     public function elementToHtml(array $Element)
@@ -77,28 +152,29 @@ trait ParsedownGravTrait
      */
     protected function blockTwigTag($Line)
     {
-        if (preg_match('/[{%|{{|{#].*[#}|}}|%}]/', $Line['body'], $matches)) {
-            $Block = array(
+        if (preg_match('/(?:{{|{%|{#)(.*)(?:}}|%}|#})/', $Line['body'], $matches)) {
+            $Block = [
                 'markup' => $Line['body'],
-            );
+            ];
+
             return $Block;
         }
     }
 
     protected function inlineSpecialCharacter($Excerpt)
     {
-        if ($Excerpt['text'][0] === '&' && ! preg_match('/^&#?\w+;/', $Excerpt['text'])) {
-            return array(
+        if ($Excerpt['text'][0] === '&' && !preg_match('/^&#?\w+;/', $Excerpt['text'])) {
+            return [
                 'markup' => '&amp;',
                 'extent' => 1,
-            );
+            ];
         }
 
         if (isset($this->special_chars[$Excerpt['text'][0]])) {
-            return array(
-                'markup' => '&'.$this->special_chars[$Excerpt['text'][0]].';',
+            return [
+                'markup' => '&' . $this->special_chars[$Excerpt['text'][0]] . ';',
                 'extent' => 1,
-            );
+            ];
         }
     }
 
@@ -109,6 +185,7 @@ trait ParsedownGravTrait
             $excerpt = parent::inlineImage($excerpt);
             $excerpt['element']['attributes']['src'] = $matches[1];
             $excerpt['extent'] = $excerpt['extent'] + strlen($matches[1]) - 1;
+
             return $excerpt;
         } else {
             $excerpt['type'] = 'image';
@@ -116,7 +193,7 @@ trait ParsedownGravTrait
         }
 
         // Some stuff we will need
-        $actions = array();
+        $actions = [];
         $media = null;
 
         // if this is an image
@@ -128,37 +205,38 @@ trait ParsedownGravTrait
             //get the url and parse it
             $url = parse_url(htmlspecialchars_decode($excerpt['element']['attributes']['src']));
 
+            $this_host = isset($url['host']) && $url['host'] == $this->uri->host();
+
             // if there is no host set but there is a path, the file is local
-            if (!isset($url['host']) && isset($url['path'])) {
+            if ((!isset($url['host']) || $this_host) && isset($url['path'])) {
                 $path_parts = pathinfo($url['path']);
 
                 // get the local path to page media if possible
                 if ($path_parts['dirname'] == $this->page->url(false, false, false)) {
-                    $url['path'] = urldecode($path_parts['basename']);
                     // get the media objects for this page
                     $media = $this->page->media();
                 } else {
                     // see if this is an external page to this one
-                    $page_route = str_replace($this->base_url, '', $path_parts['dirname']);
+                    $base_url = rtrim(self::getGrav()['base_url_relative'] . self::getGrav()['pages']->base(), '/');
+                    $page_route = '/' . ltrim(str_replace($base_url, '', $path_parts['dirname']), '/');
 
                     $ext_page = $this->pages->dispatch($page_route, true);
                     if ($ext_page) {
                         $media = $ext_page->media();
-                        $url['path'] = urldecode($path_parts['basename']);
                     }
                 }
 
                 // if there is a media file that matches the path referenced..
-                if ($media && isset($media->all()[$url['path']])) {
+                if ($media && isset($media->all()[$path_parts['basename']])) {
                     // get the medium object
-                    $medium = $media->all()[$url['path']];
+                    $medium = $media->all()[$path_parts['basename']];
 
                     // if there is a query, then parse it and build action calls
                     if (isset($url['query'])) {
                         $actions = array_reduce(explode('&', $url['query']), function ($carry, $item) {
                             $parts = explode('=', $item, 2);
                             $value = isset($parts[1]) ? $parts[1] : null;
-                            $carry[] = [ 'method' => $parts[0], 'params' => $value ];
+                            $carry[] = ['method' => $parts[0], 'params' => $value];
 
                             return $carry;
                         }, []);
@@ -166,14 +244,15 @@ trait ParsedownGravTrait
 
                     // loop through actions for the image and call them
                     foreach ($actions as $action) {
-                        $medium = call_user_func_array(array($medium, $action['method']), explode(',', urldecode($action['params'])));
+                        $medium = call_user_func_array([$medium, $action['method']],
+                            explode(',', urldecode($action['params'])));
                     }
 
                     if (isset($url['fragment'])) {
                         $medium->urlHash($url['fragment']);
                     }
 
-                    $excerpt['element'] = $medium->parseDownElement($title, $alt, $class);
+                    $excerpt['element'] = $medium->parseDownElement($title, $alt, $class, true);
 
                 } else {
                     // not a current page media file, see if it needs converting to relative
@@ -199,6 +278,7 @@ trait ParsedownGravTrait
             $excerpt = parent::inlineLink($excerpt);
             $excerpt['element']['attributes']['href'] = $matches[1];
             $excerpt['extent'] = $excerpt['extent'] + strlen($matches[1]) - 1;
+
             return $excerpt;
         } else {
             $excerpt = parent::inlineLink($excerpt);
@@ -208,13 +288,73 @@ trait ParsedownGravTrait
         if (isset($excerpt['element']['attributes']['href'])) {
             $url = parse_url(htmlspecialchars_decode($excerpt['element']['attributes']['href']));
 
-            // if there is no scheme, the file is local
-            if (!isset($url['scheme']) && (count($url) > 0)) {
-                // convert the URl is required
-                $excerpt['element']['attributes']['href'] = Uri::convertUrl($this->page, Uri::buildUrl($url), $type);
+            // if there is a query, then parse it and build action calls
+            if (isset($url['query'])) {
+                $actions = array_reduce(explode('&', $url['query']), function ($carry, $item) {
+                    $parts = explode('=', $item, 2);
+                    $value = isset($parts[1]) ? $parts[1] : true;
+                    $carry[$parts[0]] = $value;
+
+                    return $carry;
+                }, []);
+
+                // valid attributes supported
+                $valid_attributes = ['rel', 'target', 'id', 'class', 'classes'];
+
+                // Unless told to not process, go through actions
+                if (array_key_exists('noprocess', $actions)) {
+                    unset($actions['noprocess']);
+                } else {
+                    // loop through actions for the image and call them
+                    foreach ($actions as $attrib => $value) {
+                        $key = $attrib;
+
+                        if (in_array($attrib, $valid_attributes)) {
+                            // support both class and classes
+                            if ($attrib == 'classes') {
+                                $attrib = 'class';
+                            }
+                            $excerpt['element']['attributes'][$attrib] = str_replace(',', ' ', $value);
+                            unset($actions[$key]);
+                        }
+                    }
+                }
+
+                $url['query'] = http_build_query($actions, null, '&', PHP_QUERY_RFC3986);
             }
+
+            // if no query elements left, unset query
+            if (empty($url['query'])) {
+                unset ($url['query']);
+            }
+
+            // set path to / if not set
+            if (empty($url['path'])) {
+                $url['path'] = '';
+            }
+
+            // if special scheme, just return
+            if(isset($url['scheme']) && in_array($url['scheme'], $this->special_protocols)) {
+                return $excerpt;
+            }
+
+            // handle paths and such
+            $url = Uri::convertUrl($this->page, $url, $type);
+
+            // build the URL from the component parts and set it on the element
+            $excerpt['element']['attributes']['href'] = Uri::buildUrl($url);
         }
 
         return $excerpt;
+    }
+
+    // For extending this class via plugins
+    public function __call($method, $args)
+    {
+        if (isset($this->$method) === true) {
+            $func = $this->$method;
+
+            return call_user_func_array($func, $args);
+        }
     }
 }
