@@ -5,6 +5,7 @@ use Exception;
 use Grav\Common\Filesystem\Folder;
 use Grav\Common\Config\Config;
 use Grav\Common\GravTrait;
+use Grav\Common\Language\Language;
 use Grav\Common\Utils;
 use Grav\Common\Cache;
 use Grav\Common\Twig;
@@ -27,7 +28,7 @@ define('PAGE_ORDER_PREFIX_REGEX', '/^[0-9]+\./u');
  * can be retrieved via public functions. Also each page can potentially contain an array of sub-pages.
  * Recursively traversing the page structure allows Grav to create navigation systems.
  *
- * @author RocketTheme
+ * @author  RocketTheme
  * @license MIT
  */
 class Page
@@ -41,6 +42,7 @@ class Page
     protected $folder;
     protected $path;
     protected $extension;
+    protected $url_extension;
 
     protected $id;
     protected $parent;
@@ -83,6 +85,9 @@ class Page
     protected $markdown_extra;
     protected $etag;
     protected $last_modified;
+    protected $home_route;
+    protected $hide_home_route;
+    protected $ssl;
 
     /**
      * @var Page Unmodified (original) version of the page. Used for copying and moving the page.
@@ -96,16 +101,19 @@ class Page
 
     /**
      * Page Object Constructor
+     *
+     * @return $this
      */
     public function __construct()
     {
         /** @var Config $config */
         $config = self::getGrav()['config'];
 
-
-        $this->taxonomy = array();
+        $this->taxonomy = [];
         $this->process = $config->get('system.pages.process');
         $this->published = true;
+
+        return $this;
     }
 
     /**
@@ -113,12 +121,18 @@ class Page
      *
      * @param  \SplFileInfo $file The file information for the .md file that the page represents
      * @param  string       $extension
+     *
+     * @return $this
      */
     public function init(\SplFileInfo $file, $extension = null)
     {
+        $config = self::getGrav()['config'];
+
+        $this->hide_home_route = $config->get('system.home.hide_in_urls', false);
+        $this->home_route = $config->get('system.home.alias');
         $this->filePath($file->getPathName());
         $this->modified($file->getMTime());
-        $this->id($this->modified().md5($this->filePath()));
+        $this->id($this->modified() . md5($this->filePath()));
         $this->routable(true);
         $this->header();
         $this->date();
@@ -128,10 +142,11 @@ class Page
         $this->modularTwig($this->slug[0] == '_');
         $this->setPublishState();
         $this->published();
+        $this->urlExtension();
 
         // some extension logic
         if (empty($extension)) {
-            $this->extension('.'.$file->getExtension());
+            $this->extension('.' . $file->getExtension());
         } else {
             $this->extension($extension);
         }
@@ -139,6 +154,8 @@ class Page
         // extract page language from page extension
         $language = trim(basename($this->extension(), 'md'), '.') ?: null;
         $this->language($language);
+
+        return $this;
     }
 
     /**
@@ -153,13 +170,15 @@ class Page
         $translatedLanguages = [];
 
         foreach ($languages as $language) {
-            $path = $this->path . DS . $this->folder . DS . $filename . '.' . $language .'.md';
+            $path = $this->path . DS . $this->folder . DS . $filename . '.' . $language . '.md';
             if (file_exists($path)) {
                 $aPage = new Page();
-                $aPage->init(new \SplFileInfo($path), $language .'.md');
+                $aPage->init(new \SplFileInfo($path), $language . '.md');
 
                 $route = isset($aPage->header()->routes['default']) ? $aPage->header()->routes['default'] : $aPage->rawRoute();
-                if (!$route) $route = $aPage->slug();
+                if (!$route) {
+                    $route = $aPage->slug();
+                }
 
                 $translatedLanguages[$language] = $route;
             }
@@ -180,7 +199,7 @@ class Page
         $untranslatedLanguages = [];
 
         foreach ($languages as $language) {
-            $path = $this->path . DS . $this->folder . DS . $filename . '.' . $language .'.md';
+            $path = $this->path . DS . $this->folder . DS . $filename . '.' . $language . '.md';
             if (!file_exists($path)) {
                 $untranslatedLanguages[] = $language;
             }
@@ -193,6 +212,7 @@ class Page
      * Gets and Sets the raw data
      *
      * @param  string $var Raw content string
+     *
      * @return Object      Raw content string
      */
     public function raw($var = null)
@@ -207,32 +227,41 @@ class Page
 
             // Reset header and content.
             $this->modified = time();
-            $this->id($this->modified().md5($this->filePath()));
+            $this->id($this->modified() . md5($this->filePath()));
             $this->header = null;
             $this->content = null;
             $this->summary = null;
         }
+
         return $file ? $file->raw() : '';
     }
 
+    /**
+     * Gets and Sets the page frontmatter
+     *
+     * @param string|null $var
+     *
+     * @return string
+     */
     public function frontmatter($var = null)
     {
 
         if ($var) {
-            $this->frontmatter = (string) $var;
+            $this->frontmatter = (string)$var;
 
             // Update also file object.
             $file = $this->file();
             if ($file) {
-                $file->frontmatter((string) $var);
+                $file->frontmatter((string)$var);
             }
 
             // Force content re-processing.
-            $this->id(time().md5($this->filePath()));
+            $this->id(time() . md5($this->filePath()));
         }
         if (!$this->frontmatter) {
             $this->header();
         }
+
         return $this->frontmatter;
     }
 
@@ -240,21 +269,22 @@ class Page
      * Gets and Sets the header based on the YAML configuration at the top of the .md file
      *
      * @param  object|array $var a YAML object representing the configuration for the file
+     *
      * @return object      the current YAML configuration
      */
     public function header($var = null)
     {
         if ($var) {
-            $this->header = (object) $var;
+            $this->header = (object)$var;
 
             // Update also file object.
             $file = $this->file();
             if ($file) {
-                $file->header((array) $var);
+                $file->header((array)$var);
             }
 
             // Force content re-processing.
-            $this->id(time().md5($this->filePath()));
+            $this->id(time() . md5($this->filePath()));
         }
         if (!$this->header) {
             $file = $this->file();
@@ -266,7 +296,13 @@ class Page
                     $this->frontmatter = $file->frontmatter();
                     $this->header = (object)$file->header();
                 } catch (ParseException $e) {
-                    $file->raw(self::getGrav()['language']->translate(['FRONTMATTER_ERROR_PAGE', $this->slug(), $file->filename(), $e->getMessage(), $file->raw()]));
+                    $file->raw(self::getGrav()['language']->translate([
+                        'FRONTMATTER_ERROR_PAGE',
+                        $this->slug(),
+                        $file->filename(),
+                        $e->getMessage(),
+                        $file->raw()
+                    ]));
                     $this->raw_content = $file->markdown();
                     $this->frontmatter = $file->frontmatter();
                     $this->header = (object)$file->header();
@@ -295,10 +331,10 @@ class Page
                 $this->menu = trim($this->header->menu);
             }
             if (isset($this->header->routable)) {
-                $this->routable = (bool) $this->header->routable;
+                $this->routable = (bool)$this->header->routable;
             }
             if (isset($this->header->visible)) {
-                $this->visible = (bool) $this->header->visible;
+                $this->visible = (bool)$this->header->visible;
             }
             if (isset($this->header->redirect)) {
                 $this->redirect = trim($this->header->redirect);
@@ -319,7 +355,7 @@ class Page
                 $this->markdown_extra = (bool)$this->header->markdown_extra;
             }
             if (isset($this->header->taxonomy)) {
-                foreach ((array) $this->header->taxonomy as $taxonomy => $taxitems) {
+                foreach ((array)$this->header->taxonomy as $taxonomy => $taxitems) {
                     $this->taxonomy[$taxonomy] = (array)$taxitems;
                 }
             }
@@ -327,12 +363,12 @@ class Page
                 $this->max_count = intval($this->header->max_count);
             }
             if (isset($this->header->process)) {
-                foreach ((array) $this->header->process as $process => $status) {
-                    $this->process[$process] = (bool) $status;
+                foreach ((array)$this->header->process as $process => $status) {
+                    $this->process[$process] = (bool)$status;
                 }
             }
             if (isset($this->header->published)) {
-                $this->published = (bool) $this->header->published;
+                $this->published = (bool)$this->header->published;
             }
             if (isset($this->header->publish_date)) {
                 $this->publishDate($this->header->publish_date);
@@ -344,12 +380,14 @@ class Page
                 $this->expires = intval($this->header->expires);
             }
             if (isset($this->header->etag)) {
-                $this->etag = (bool) $this->header->etag;
+                $this->etag = (bool)$this->header->etag;
             }
             if (isset($this->header->last_modified)) {
-                $this->last_modified = (bool) $this->header->last_modified;
+                $this->last_modified = (bool)$this->header->last_modified;
             }
-
+            if (isset($this->header->ssl)) {
+                $this->ssl = (bool) $this->header->ssl;
+            }
         }
 
         return $this->header;
@@ -379,19 +417,19 @@ class Page
      */
     public function modifyHeader($key, $value)
     {
-        $this->header->$key = $value;
+        $this->header->{$key} = $value;
     }
 
     /**
      * Get the summary.
      *
-     * @param  int    $size Max summary size.
+     * @param  int $size Max summary size.
+     *
      * @return string
      */
     public function summary($size = null)
     {
-        /** @var Config $config */
-        $config = self::getGrav()['config']->get('site.summary');
+        $config = (array)self::getGrav()['config']->get('site.summary');
         if (isset($this->header->summary)) {
             $config = array_merge($config, $this->header->summary);
         }
@@ -413,7 +451,7 @@ class Page
         // Return calculated summary based on summary divider's position
         $format = $config['format'];
         // Return entire page content on wrong/ unknown format
-        if (!in_array($format, array('short', 'long'))) {
+        if (!in_array($format, ['short', 'long'])) {
             return $content;
         } elseif (($format === 'short') && isset($summary_size)) {
             return mb_substr($content, 0, $summary_size);
@@ -427,7 +465,7 @@ class Page
         // If the size is zero, return the entire page content
         if ($size === 0) {
             return $content;
-        // Return calculated summary based on defaults
+            // Return calculated summary based on defaults
         } elseif (!is_numeric($size) || ($size < 0)) {
             $size = 300;
         }
@@ -438,7 +476,7 @@ class Page
     /**
      * Sets the summary of the page
      *
-     * @param string $var Summary
+     * @param string $summary Summary
      */
     public function setSummary($summary)
     {
@@ -449,6 +487,7 @@ class Page
      * Gets and Sets the content based on content portion of the .md file
      *
      * @param  string $var Content
+     *
      * @return string      Content
      */
     public function content($var = null)
@@ -463,7 +502,7 @@ class Page
             }
 
             // Force re-processing.
-            $this->id(time().md5($this->filePath()));
+            $this->id(time() . md5($this->filePath()));
             $this->content = null;
         }
         // If no content, process it
@@ -471,59 +510,59 @@ class Page
             // Get media
             $this->media();
 
+            /** @var Config $config */
+            $config = self::getGrav()['config'];
+
             // Load cached content
             /** @var Cache $cache */
             $cache = self::getGrav()['cache'];
-            $cache_id = md5('page'.$this->id());
+            $cache_id = md5('page' . $this->id());
             $this->content = $cache->fetch($cache_id);
 
             $process_markdown = $this->shouldProcess('markdown');
             $process_twig = $this->shouldProcess('twig');
-            $cache_twig = isset($this->header->cache_enable) ? $this->header->cache_enable : true;
-            $twig_first = isset($this->header->twig_first) ? $this->header->twig_first : false;
-            $twig_already_processed = false;
+            $cache_enable = isset($this->header->cache_enable) ? $this->header->cache_enable : $config->get('system.cache.enabled',
+                true);
+            $twig_first = isset($this->header->twig_first) ? $this->header->twig_first : $config->get('system.pages.twig_first',
+                true);
+
 
             // if no cached-content run everything
-            if ($this->content === false) {
+            if ($this->content === false || $cache_enable === false) {
                 $this->content = $this->raw_content;
                 self::getGrav()->fireEvent('onPageContentRaw', new Event(['page' => $this]));
 
                 if ($twig_first) {
                     if ($process_twig) {
                         $this->processTwig();
-                        $twig_already_processed = true;
                     }
                     if ($process_markdown) {
                         $this->processMarkdown();
                     }
-                    if ($cache_twig) {
-                        $this->cachePageContent();
-                    }
+
+                    // Content Processed but not cached yet
+                    self::getGrav()->fireEvent('onPageContentProcessed', new Event(['page' => $this]));
+
                 } else {
                     if ($process_markdown) {
                         $this->processMarkdown();
                     }
-                    if (!$cache_twig) {
-                        $this->cachePageContent();
-                    }
+
+                    // Content Processed but not cached yet
+                    self::getGrav()->fireEvent('onPageContentProcessed', new Event(['page' => $this]));
+
                     if ($process_twig) {
                         $this->processTwig();
-                        $twig_already_processed = true;
-                    }
-                    if ($cache_twig) {
-                        $this->cachePageContent();
                     }
                 }
-            // content cached, but twig cache off
-            }
 
-            // only markdown content cached, process twig if required and not already processed
-            if ($process_twig && !$cache_twig && !$twig_already_processed) {
-                $this->processTwig();
+                if ($cache_enable) {
+                    $this->cachePageContent();
+                }
             }
 
             // Handle summary divider
-            $delimiter = self::getGrav()['config']->get('site.summary.delimiter', '===');
+            $delimiter = $config->get('site.summary.delimiter', '===');
             $divider_pos = mb_strpos($this->content, "<p>{$delimiter}</p>");
             if ($divider_pos !== false) {
                 $this->summary_size = $divider_pos;
@@ -543,7 +582,7 @@ class Page
         /** @var Config $config */
         $config = self::getGrav()['config'];
 
-        $defaults = (array) $config->get('system.pages.markdown');
+        $defaults = (array)$config->get('system.pages.markdown');
         if (isset($this->header()->markdown)) {
             $defaults = array_merge($defaults, $this->header()->markdown);
         }
@@ -579,9 +618,7 @@ class Page
     private function cachePageContent()
     {
         $cache = self::getGrav()['cache'];
-        $cache_id = md5('page'.$this->id());
-
-        self::getGrav()->fireEvent('onPageContentProcessed', new Event(['page' => $this]));
+        $cache_id = md5('page' . $this->id());
         $cache->save($cache_id, $this->content);
     }
 
@@ -608,8 +645,9 @@ class Page
     /**
      * Get value from a page variable (used mostly for creating edit forms).
      *
-     * @param  string  $name  Variable name.
-     * @param mixed $default
+     * @param string $name Variable name.
+     * @param mixed  $default
+     *
      * @return mixed
      */
     public function value($name, $default = null)
@@ -622,21 +660,24 @@ class Page
         }
         if ($name == 'order') {
             $order = $this->order();
-            return $order ? (int) $this->order() : '';
+
+            return $order ? (int)$this->order() : '';
         }
         if ($name == 'ordering') {
-            return (bool) $this->order();
+            return (bool)$this->order();
         }
         if ($name == 'folder') {
             $regex = '/^[0-9]+\./u';
+
             return preg_replace($regex, '', $this->folder);
         }
         if ($name == 'name') {
             $language = $this->language() ? '.' . $this->language() : '';
-            $name_val = str_replace($language .'.md', '', $this->name());
+            $name_val = str_replace($language . '.md', '', $this->name());
             if ($this->modular()) {
                 return 'modular/' . $name_val;
             }
+
             return $name_val;
         }
         if ($name == 'media') {
@@ -680,6 +721,13 @@ class Page
         return $default;
     }
 
+    /**
+     * Gets and Sets the Page raw content
+     *
+     * @param null $var
+     *
+     * @return null
+     */
     public function rawMarkdown($var = null)
     {
         if ($var !== null) {
@@ -699,11 +747,13 @@ class Page
         if ($this->name) {
             return MarkdownFile::instance($this->filePath());
         }
+
         return null;
     }
 
     /**
      * Save page if there's a file assigned to it.
+     *
      * @param bool $reorder Internal use.
      */
     public function save($reorder = true)
@@ -714,7 +764,7 @@ class Page
         $file = $this->file();
         if ($file) {
             $file->filename($this->filePath());
-            $file->header((array) $this->header());
+            $file->header((array)$this->header());
             $file->markdown($this->raw_content);
             $file->save();
         }
@@ -726,27 +776,31 @@ class Page
      * You need to call $this->save() in order to perform the move.
      *
      * @param Page $parent New parent page.
-     * @return Page
+     *
+     * @return $this
      */
     public function move(Page $parent)
     {
-        $clone = clone $this;
-        $clone->_action = 'move';
-        $clone->_original = $this;
-        $clone->parent($parent);
-        $clone->id(time().md5($clone->filePath()));
+        if (!$this->_original) {
+            $clone = clone $this;
+            $this->_original = $clone;
+        }
+
+        $this->_action = 'move';
+        $this->parent($parent);
+        $this->id(time() . md5($this->filePath()));
 
         if ($parent->path()) {
-            $clone->path($parent->path() . '/' . $clone->folder());
+            $this->path($parent->path() . '/' . $this->folder());
         }
 
         if ($parent->route()) {
-            $clone->route($parent->route() . '/'. $clone->slug());
+            $this->route($parent->route() . '/' . $this->slug());
         } else {
-            $clone->route(self::getGrav()['pages']->root()->route() . '/'. $clone->slug());
+            $this->route(self::getGrav()['pages']->root()->route() . '/' . $this->slug());
         }
 
-        return $clone;
+        return $this;
     }
 
     /**
@@ -756,14 +810,15 @@ class Page
      * You need to call $this->save() in order to perform the move.
      *
      * @param Page $parent New parent page.
-     * @return Page
+     *
+     * @return $this
      */
     public function copy($parent)
     {
-        $clone = $this->move($parent);
-        $clone->_action = 'copy';
+        $this->move($parent);
+        $this->_action = 'copy';
 
-        return $clone;
+        return $this;
     }
 
     /**
@@ -824,7 +879,7 @@ class Page
         $blueprints = $this->blueprints();
         $values = $blueprints->filter($this->toArray());
         if ($values && isset($values['header'])) {
-        	$this->header($values['header']);
+            $this->header($values['header']);
         }
     }
 
@@ -836,6 +891,7 @@ class Page
     public function extra()
     {
         $blueprints = $this->blueprints();
+
         return $blueprints->extra($this->toArray()['header'], 'header.');
     }
 
@@ -846,10 +902,10 @@ class Page
      */
     public function toArray()
     {
-        return array(
-            'header' => (array) $this->header(),
-            'content' => (string) $this->value('content')
-        );
+        return [
+            'header'  => (array)$this->header(),
+            'content' => (string)$this->value('content')
+        ];
     }
 
     /**
@@ -876,6 +932,7 @@ class Page
      * Gets and sets the associated media as found in the page folder.
      *
      * @param  Media $var Representation of associated media.
+     *
      * @return Media      Representation of associated media.
      */
     public function media($var = null)
@@ -888,13 +945,14 @@ class Page
         }
         if ($this->media === null) {
             // Use cached media if possible.
-            $media_cache_id = md5('media'.$this->id());
+            $media_cache_id = md5('media' . $this->id());
             if (!$media = $cache->fetch($media_cache_id)) {
                 $media = new Media($this->path());
                 $cache->save($media_cache_id, $media);
             }
             $this->media = $media;
         }
+
         return $this->media;
     }
 
@@ -902,6 +960,7 @@ class Page
      * Gets and sets the name field.  If no name field is set, it will return 'default.md'.
      *
      * @param  string $var The name of this page.
+     *
      * @return string      The name of this page.
      */
     public function name($var = null)
@@ -909,6 +968,7 @@ class Page
         if ($var !== null) {
             $this->name = $var;
         }
+
         return empty($this->name) ? 'default.md' : $this->name;
     }
 
@@ -919,7 +979,7 @@ class Page
      */
     public function childType()
     {
-        return isset($this->header->child_type) ? (string) $this->header->child_type : 'default';
+        return isset($this->header->child_type) ? (string)$this->header->child_type : 'default';
     }
 
     /**
@@ -927,6 +987,7 @@ class Page
      * If no field is set, it will return the name without the .md extension
      *
      * @param  string $var the template name
+     *
      * @return string      the template name
      */
     public function template($var = null)
@@ -937,6 +998,7 @@ class Page
         if (empty($this->template)) {
             $this->template = ($this->modular() ? 'modular/' : '') . str_replace($this->extension(), '', $this->name());
         }
+
         return $this->template;
     }
 
@@ -944,6 +1006,7 @@ class Page
      * Gets and sets the extension field.
      *
      * @param null $var
+     *
      * @return null|string
      */
     public function extension($var = null)
@@ -954,20 +1017,43 @@ class Page
         if (empty($this->extension)) {
             $this->extension = '.' . pathinfo($this->name(), PATHINFO_EXTENSION);
         }
+
         return $this->extension;
+    }
+
+    /**
+     * Returns the page extension, got from the page `url_extension` config and falls back to the
+     * system config `system.pages.append_url_extension`.
+     *
+     * @return string      The extension of this page. For example `.html`
+     */
+    public function urlExtension()
+    {
+        if ($this->home()) {
+            return '';
+        }
+
+        // if not set in the page get the value from system config
+        if (empty($this->url_extension)) {
+            $this->url_extension = trim(isset($this->header->append_url_extension) ? $this->header->append_url_extension : self::getGrav()['config']->get('system.pages.append_url_extension', false));
+        }
+
+        return $this->url_extension;
     }
 
     /**
      * Gets and sets the expires field. If not set will return the default
      *
-     * @param  string $var The name of this page.
-     * @return string      The name of this page.
+     * @param  int $var The new expires value.
+     *
+     * @return int      The expires value
      */
     public function expires($var = null)
     {
         if ($var !== null) {
             $this->expires = $var;
         }
+
         return empty($this->expires) ? self::getGrav()['config']->get('system.pages.expires') : $this->expires;
     }
 
@@ -975,6 +1061,7 @@ class Page
      * Gets and sets the title for this Page.  If no title is set, it will use the slug() to get a name
      *
      * @param  string $var the title of the Page
+     *
      * @return string      the title of the Page
      */
     public function title($var = null)
@@ -985,6 +1072,7 @@ class Page
         if (empty($this->title)) {
             $this->title = ucfirst($this->slug());
         }
+
         return $this->title;
     }
 
@@ -993,6 +1081,7 @@ class Page
      * If no menu field is set, it will use the title()
      *
      * @param  string $var the menu field for the page
+     *
      * @return string      the menu field for the page
      */
     public function menu($var = null)
@@ -1003,6 +1092,7 @@ class Page
         if (empty($this->menu)) {
             $this->menu = $this->title();
         }
+
         return $this->menu;
     }
 
@@ -1010,12 +1100,13 @@ class Page
      * Gets and Sets whether or not this Page is visible for navigation
      *
      * @param  bool $var true if the page is visible
+     *
      * @return bool      true if the page is visible
      */
     public function visible($var = null)
     {
         if ($var !== null) {
-            $this->visible = (bool) $var;
+            $this->visible = (bool)$var;
         }
 
         if ($this->visible === null) {
@@ -1028,6 +1119,7 @@ class Page
                 $this->visible = false;
             }
         }
+
         return $this->visible;
     }
 
@@ -1035,12 +1127,13 @@ class Page
      * Gets and Sets whether or not this Page is considered published
      *
      * @param  bool $var true if the page is published
+     *
      * @return bool      true if the page is published
      */
     public function published($var = null)
     {
         if ($var !== null) {
-            $this->published = (bool) $var;
+            $this->published = (bool)$var;
         }
 
         // If not published, should not be visible in menus either
@@ -1055,16 +1148,13 @@ class Page
      * Gets and Sets the Page publish date
      *
      * @param  string $var string representation of a date
+     *
      * @return int         unix timestamp representation of the date
      */
     public function publishDate($var = null)
     {
         if ($var !== null) {
             $this->publish_date = Utils::date2timestamp($var);
-        }
-
-        if ($this->publish_date === null) {
-            $this->publish_date = $this->date();
         }
 
         return $this->publish_date;
@@ -1074,6 +1164,7 @@ class Page
      * Gets and Sets the Page unpublish date
      *
      * @param  string $var string representation of a date
+     *
      * @return int|null         unix timestamp representation of the date
      */
     public function unpublishDate($var = null)
@@ -1091,15 +1182,25 @@ class Page
      * The page must be *routable* and *published*
      *
      * @param  bool $var true if the page is routable
+     *
      * @return bool      true if the page is routable
      */
     public function routable($var = null)
     {
         if ($var !== null) {
-            $this->routable = (bool) $var;
+            $this->routable = (bool)$var;
         }
 
         return $this->routable && $this->published();
+    }
+
+    public function ssl($var = null)
+    {
+        if ($var !== null) {
+            $this->ssl = (bool) $var;
+        }
+
+        return $this->ssl;
     }
 
     /**
@@ -1107,13 +1208,15 @@ class Page
      * a simple array of arrays with the form array("markdown"=>true) for example
      *
      * @param  array $var an Array of name value pairs where the name is the process and value is true or false
+     *
      * @return array      an Array of name value pairs where the name is the process and value is true or false
      */
     public function process($var = null)
     {
         if ($var !== null) {
-            $this->process = (array) $var;
+            $this->process = (array)$var;
         }
+
         return $this->process;
     }
 
@@ -1122,12 +1225,13 @@ class Page
      * that can then be rendered in the page.
      *
      * @param  array $var an Array of metadata values to set
+     *
      * @return array      an Array of metadata values for the page
      */
     public function metadata($var = null)
     {
         if ($var !== null) {
-            $this->metadata = (array) $var;
+            $this->metadata = (array)$var;
         }
 
         // if not metadata yet, process it.
@@ -1136,11 +1240,12 @@ class Page
 
             $this->metadata = [];
 
+            $metadata = [];
             // Set the Generator tag
-            $this->metadata['generator'] = array('name'=>'generator', 'content'=>'GravCMS ' . GRAV_VERSION);
+            $metadata['generator'] = 'GravCMS';
 
             // Get initial metadata for the page
-            $metadata  = self::getGrav()['config']->get('site.metadata');
+            $metadata = array_merge($metadata, self::getGrav()['config']->get('site.metadata'));
 
             if (isset($this->header->metadata)) {
                 // Merge any site.metadata settings in with page metadata
@@ -1150,17 +1255,29 @@ class Page
             // Build an array of meta objects..
             foreach ((array)$metadata as $key => $value) {
                 // If this is a property type metadata: "og", "twitter", "facebook" etc
+                // Backward compatibility for nested arrays in metas
                 if (is_array($value)) {
                     foreach ($value as $property => $prop_value) {
-                        $prop_key =  $key.":".$property;
-                        $this->metadata[$prop_key] = array('property'=>$prop_key, 'content'=>htmlspecialchars($prop_value, ENT_QUOTES));
+                        $prop_key                  = $key . ":" . $property;
+                        $this->metadata[$prop_key] = ['name' => $prop_key, 'property' => $prop_key, 'content' => htmlspecialchars($prop_value, ENT_QUOTES)];
                     }
-                // If it this is a standard meta data type
                 } else {
-                    if (in_array($key, $header_tag_http_equivs)) {
-                        $this->metadata[$key] = array('http_equiv'=>$key, 'content'=>htmlspecialchars($value, ENT_QUOTES));
-                    } else {
-                        $this->metadata[$key] = array('name'=>$key, 'content'=>htmlspecialchars($value, ENT_QUOTES));
+                    // If it this is a standard meta data type
+                    if ($value) {
+                        if (in_array($key, $header_tag_http_equivs)) {
+                            $this->metadata[$key] = ['http_equiv' => $key, 'content' => htmlspecialchars($value, ENT_QUOTES)];
+                        } else {
+                            // if it's a social metadata with separator, render as property
+                            $separator    = strpos($key, ':');
+                            $hasSeparator = $separator && $separator < strlen($key) - 1;
+                            $entry        = ['name' => $key, 'content' => htmlspecialchars($value, ENT_QUOTES)];
+
+                            if ($hasSeparator) {
+                                $entry['property'] = $key;
+                            }
+
+                            $this->metadata[$key] = $entry;
+                        }
                     }
                 }
             }
@@ -1174,6 +1291,7 @@ class Page
      * the parent folder from the path
      *
      * @param  string $var the slug, e.g. 'my-blog'
+     *
      * @return string      the slug
      */
     public function slug($var = null)
@@ -1193,6 +1311,7 @@ class Page
      * Get/set order number of this page.
      *
      * @param int $var
+     *
      * @return int|bool
      */
     public function order($var = null)
@@ -1219,6 +1338,7 @@ class Page
      * Gets the URL for a page - alias of url().
      *
      * @param bool $include_host
+     *
      * @return string the permalink
      */
     public function link($include_host = false)
@@ -1229,25 +1349,36 @@ class Page
     /**
      * Gets the url for the Page.
      *
-     * @param  bool $include_host Defaults false, but true would include http://yourhost.com
-     * @param  bool $canonical true to return the canonical URL
+     * @param bool $include_host Defaults false, but true would include http://yourhost.com
+     * @param bool $canonical    true to return the canonical URL
+     * @param bool $include_lang
      *
      * @return string The url.
      */
     public function url($include_host = false, $canonical = false, $include_lang = true)
     {
-
         /** @var Pages $pages */
         $pages = self::getGrav()['pages'];
+
+        /** @var Config $config */
+        $config = self::getGrav()['config'];
 
         /** @var Language $language */
         $language = self::getGrav()['language'];
 
+        /** @var Uri $uri */
+        $uri = self::getGrav()['uri'];
+
         // get pre-route
         if ($include_lang && $language->enabled()) {
-           $pre_route = $language->getLanguageURLPrefix();
+            $pre_route = $language->getLanguageURLPrefix();
         } else {
             $pre_route = '';
+        }
+
+        // add full route if configured to do so
+        if ($config->get('system.absolute_urls', false)) {
+            $include_host = true;
         }
 
         // get canonical route if requested
@@ -1257,12 +1388,9 @@ class Page
             $route = $pre_route . $this->route();
         }
 
-        /** @var Uri $uri */
-        $uri = self::getGrav()['uri'];
-
         $rootUrl = $uri->rootUrl($include_host) . $pages->base();
 
-        $url = $rootUrl.'/'.trim($route, '/');
+        $url = $rootUrl . '/' . trim($route, '/') . $this->urlExtension();
 
         // trim trailing / if not root
         if ($url !== '/') {
@@ -1276,7 +1404,7 @@ class Page
      * Gets the route for the page based on the route headers if available, else from
      * the parents route and the current Page's slug.
      *
-     * @param  string  $var  Set new default route.
+     * @param  string $var Set new default route.
      *
      * @return string  The route for the Page.
      */
@@ -1287,13 +1415,24 @@ class Page
         }
 
         if (empty($this->route)) {
+            $baseRoute = null;
+
             // calculate route based on parent slugs
-            $baseRoute = $this->parent ? (string) $this->parent()->route() : null;
-            $this->route = isset($baseRoute) ? $baseRoute . '/'. $this->slug() : null;
+            $parent = $this->parent();
+            if (isset($parent)) {
+                if ($this->hide_home_route && $parent->route() == $this->home_route) {
+                    $baseRoute = '';
+                } else {
+                    $baseRoute = (string)$parent->route();
+                }
+            }
+
+            $this->route = isset($baseRoute) ? $baseRoute . '/' . $this->slug() : null;
 
             if (!empty($this->routes) && isset($this->routes['default'])) {
                 $this->routes['aliases'][] = $this->route;
                 $this->route = $this->routes['default'];
+
                 return $this->route;
             }
         }
@@ -1310,6 +1449,13 @@ class Page
         unset($this->slug);
     }
 
+    /**
+     * Gets and Sets the page raw route
+     *
+     * @param null $var
+     *
+     * @return null|string
+     */
     public function rawRoute($var = null)
     {
         if ($var !== null) {
@@ -1317,12 +1463,12 @@ class Page
         }
 
         if (empty($this->raw_route)) {
-            $baseRoute = $this->parent ? (string) $this->parent()->rawRoute() : null;
+            $baseRoute = $this->parent ? (string)$this->parent()->rawRoute() : null;
 
             $regex = '/^[0-9]+\./u';
             $slug = preg_replace($regex, '', $this->folder);
 
-            $this->raw_route = isset($baseRoute) ? $baseRoute . '/'. $slug : null;
+            $this->raw_route = isset($baseRoute) ? $baseRoute . '/' . $slug : null;
         }
 
         return $this->raw_route;
@@ -1331,14 +1477,14 @@ class Page
     /**
      * Gets the route aliases for the page based on page headers.
      *
-     * @param  array  $var  list of route aliases
+     * @param  array $var list of route aliases
      *
      * @return array  The route aliases for the Page.
      */
     public function routeAliases($var = null)
     {
         if ($var !== null) {
-            $this->routes['aliases'] = (array) $var;
+            $this->routes['aliases'] = (array)$var;
         }
 
         if (!empty($this->routes) && isset($this->routes['aliases'])) {
@@ -1373,6 +1519,7 @@ class Page
      * Gets and sets the identifier for this Page object.
      *
      * @param  string $var the identifier
+     *
      * @return string      the identifier
      */
     public function id($var = null)
@@ -1380,6 +1527,7 @@ class Page
         if ($var !== null) {
             $this->id = $var;
         }
+
         return $this->id;
     }
 
@@ -1387,6 +1535,7 @@ class Page
      * Gets and sets the modified timestamp.
      *
      * @param  int $var modified unix timestamp
+     *
      * @return int      modified unix timestamp
      */
     public function modified($var = null)
@@ -1394,6 +1543,7 @@ class Page
         if ($var !== null) {
             $this->modified = $var;
         }
+
         return $this->modified;
     }
 
@@ -1401,6 +1551,7 @@ class Page
      * Gets the redirect set in the header.
      *
      * @param  string $var redirect url
+     *
      * @return array
      */
     public function redirect($var = null)
@@ -1408,6 +1559,7 @@ class Page
         if ($var !== null) {
             $this->redirect = $var;
         }
+
         return $this->redirect;
     }
 
@@ -1415,6 +1567,7 @@ class Page
      * Gets and sets the option to show the etag header for the page.
      *
      * @param  boolean $var show etag header
+     *
      * @return boolean      show etag header
      */
     public function eTag($var = null)
@@ -1423,8 +1576,9 @@ class Page
             $this->etag = $var;
         }
         if (!isset($this->etag)) {
-            $this->etag = (bool) self::getGrav()['config']->get('system.pages.etag');
+            $this->etag = (bool)self::getGrav()['config']->get('system.pages.etag');
         }
+
         return $this->etag;
     }
 
@@ -1432,6 +1586,7 @@ class Page
      * Gets and sets the option to show the last_modified header for the page.
      *
      * @param  boolean $var show last_modified header
+     *
      * @return boolean      show last_modified header
      */
     public function lastModified($var = null)
@@ -1440,8 +1595,9 @@ class Page
             $this->last_modified = $var;
         }
         if (!isset($this->last_modified)) {
-            $this->last_modified = (bool) self::getGrav()['config']->get('system.pages.last_modified');
+            $this->last_modified = (bool)self::getGrav()['config']->get('system.pages.last_modified');
         }
+
         return $this->last_modified;
     }
 
@@ -1449,6 +1605,7 @@ class Page
      * Gets and sets the path to the .md file for this Page object.
      *
      * @param  string $var the file path
+     *
      * @return string|null      the file path
      */
     public function filePath($var = null)
@@ -1461,6 +1618,7 @@ class Page
             // Path to the page.
             $this->path = dirname(dirname($var));
         }
+
         return $this->path . '/' . $this->folder . '/' . ($this->name ?: '');
     }
 
@@ -1471,7 +1629,19 @@ class Page
      */
     public function filePathClean()
     {
-        return str_replace(ROOT_DIR, '', $this->filePath());
+        $path = str_replace(ROOT_DIR, '', $this->filePath());
+
+        return $path;
+    }
+
+    /**
+     * Returns the clean path to the page file
+     */
+    public function relativePagePath()
+    {
+        $path = str_replace('/' . $this->name, '', $this->filePathClean());
+
+        return $path;
     }
 
     /**
@@ -1479,6 +1649,7 @@ class Page
      * This is equivalent to the filePath but without the filename.
      *
      * @param  string $var the path
+     *
      * @return string|null      the path
      */
     public function path($var = null)
@@ -1489,6 +1660,7 @@ class Page
             // Path to the page.
             $this->path = dirname($var);
         }
+
         return $this->path ? $this->path . '/' . $this->folder : null;
     }
 
@@ -1496,6 +1668,7 @@ class Page
      * Get/set the folder.
      *
      * @param string $var Optional path
+     *
      * @return string|null
      */
     public function folder($var = null)
@@ -1503,6 +1676,7 @@ class Page
         if ($var !== null) {
             $this->folder = $var;
         }
+
         return $this->folder;
     }
 
@@ -1510,6 +1684,7 @@ class Page
      * Gets and sets the date for this Page object. This is typically passed in via the page headers
      *
      * @param  string $var string representation of a date
+     *
      * @return int         unix timestamp representation of the date
      */
     public function date($var = null)
@@ -1527,7 +1702,9 @@ class Page
 
     /**
      * Gets and sets the order by which any sub-pages should be sorted.
+     *
      * @param  string $var the order, either "asc" or "desc"
+     *
      * @return string      the order, either "asc" or "desc"
      */
     public function orderDir($var = null)
@@ -1538,6 +1715,7 @@ class Page
         if (empty($this->order_dir)) {
             $this->order_dir = 'asc';
         }
+
         return $this->order_dir;
     }
 
@@ -1550,6 +1728,7 @@ class Page
      * folder - is the order based on the name of the folder with any numerics omitted
      *
      * @param  string $var supported options include "default", "title", "date", and "folder"
+     *
      * @return string      supported options include "default", "title", "date", and "folder"
      */
     public function orderBy($var = null)
@@ -1557,6 +1736,7 @@ class Page
         if ($var !== null) {
             $this->order_by = $var;
         }
+
         return $this->order_by;
     }
 
@@ -1564,6 +1744,7 @@ class Page
      * Gets the manual order set in the header.
      *
      * @param  string $var supported options include "default", "title", "date", and "folder"
+     *
      * @return array
      */
     public function orderManual($var = null)
@@ -1571,7 +1752,8 @@ class Page
         if ($var !== null) {
             $this->order_manual = $var;
         }
-        return (array) $this->order_manual;
+
+        return (array)$this->order_manual;
     }
 
     /**
@@ -1579,18 +1761,20 @@ class Page
      * sub_pages header property is set for this page object.
      *
      * @param  int $var the maximum number of sub-pages
+     *
      * @return int      the maximum number of sub-pages
      */
     public function maxCount($var = null)
     {
         if ($var !== null) {
-            $this->max_count = (int) $var;
+            $this->max_count = (int)$var;
         }
         if (empty($this->max_count)) {
             /** @var Config $config */
             $config = self::getGrav()['config'];
-            $this->max_count = (int) $config->get('system.pages.list.count');
+            $this->max_count = (int)$config->get('system.pages.list.count');
         }
+
         return $this->max_count;
     }
 
@@ -1598,6 +1782,7 @@ class Page
      * Gets and sets the taxonomy array which defines which taxonomies this page identifies itself with.
      *
      * @param  array $var an array of taxonomies
+     *
      * @return array      an array of taxonomies
      */
     public function taxonomy($var = null)
@@ -1605,13 +1790,15 @@ class Page
         if ($var !== null) {
             $this->taxonomy = $var;
         }
+
         return $this->taxonomy;
     }
 
-     /**
+    /**
      * Gets and sets the modular var that helps identify this parent page contains modular pages.
      *
      * @param  bool $var true if modular_twig
+     *
      * @return bool      true if modular_twig
      */
     public function modular($var = null)
@@ -1624,12 +1811,13 @@ class Page
      * twig processing handled differently from a regular page.
      *
      * @param  bool $var true if modular_twig
+     *
      * @return bool      true if modular_twig
      */
     public function modularTwig($var = null)
     {
         if ($var !== null) {
-            $this->modular_twig = (bool) $var;
+            $this->modular_twig = (bool)$var;
             if ($var) {
                 $this->process['twig'] = true;
                 $this->visible(false);
@@ -1639,6 +1827,7 @@ class Page
                 }
             }
         }
+
         return $this->modular_twig;
     }
 
@@ -1646,23 +1835,26 @@ class Page
      * Gets the configured state of the processing method.
      *
      * @param  string $process the process, eg "twig" or "markdown"
+     *
      * @return bool            whether or not the processing method is enabled for this Page
      */
     public function shouldProcess($process)
     {
-        return isset($this->process[$process]) ? (bool) $this->process[$process] : false;
+        return isset($this->process[$process]) ? (bool)$this->process[$process] : false;
     }
 
     /**
      * Gets and Sets the parent object for this page
      *
      * @param  Page $var the parent page object
+     *
      * @return Page|null the parent page object if it exists.
      */
     public function parent(Page $var = null)
     {
         if ($var) {
             $this->parent = $var->path();
+
             return $var;
         }
 
@@ -1670,6 +1862,31 @@ class Page
         $pages = self::getGrav()['pages'];
 
         return $pages->get($this->parent);
+    }
+
+    /**
+     * Gets the top parent object for this page
+     *
+     * @return Page|null the top parent page object if it exists.
+     */
+    public function topParent()
+    {
+        $topParent = $this->parent();
+
+        if (!$topParent) {
+            return null;
+        }
+
+        while (true) {
+            $theParent = $topParent->parent();
+            if ($theParent !== null && $theParent->parent() !== null) {
+                $topParent = $theParent;
+            } else {
+                break;
+            }
+        }
+
+        return $topParent;
     }
 
     /**
@@ -1681,6 +1898,7 @@ class Page
     {
         /** @var Pages $pages */
         $pages = self::getGrav()['pages'];
+
         return $pages->children($this->path());
     }
 
@@ -1696,6 +1914,7 @@ class Page
         if ($collection instanceof Collection) {
             return $collection->isFirst($this->path());
         }
+
         return true;
     }
 
@@ -1710,6 +1929,7 @@ class Page
         if ($collection instanceof Collection) {
             return $collection->isLast($this->path());
         }
+
         return true;
     }
 
@@ -1737,6 +1957,7 @@ class Page
      * Returns the adjacent sibling based on a direction.
      *
      * @param  integer $direction either -1 or +1
+     *
      * @return Page             the sibling page
      */
     public function adjacentSibling($direction = 1)
@@ -1745,6 +1966,7 @@ class Page
         if ($collection instanceof Collection) {
             return $collection->adjacentSibling($this->path(), $direction);
         }
+
         return false;
     }
 
@@ -1764,6 +1986,7 @@ class Page
             }
 
         }
+
         return false;
     }
 
@@ -1781,12 +2004,15 @@ class Page
         $routes = self::getGrav()['pages']->routes();
 
         if (isset($routes[$uri_path])) {
+            /** @var Page $child_page */
             $child_page = $pages->dispatch($uri->route())->parent();
-            if ($child_page) while (!$child_page->root()) {
-                if ($this->path() == $child_page->path()) {
-                    return true;
+            if ($child_page) {
+                while (!$child_page->root()) {
+                    if ($this->path() == $child_page->path()) {
+                        return true;
+                    }
+                    $child_page = $child_page->parent();
                 }
-                $child_page = $child_page->parent();
             }
         }
 
@@ -1820,8 +2046,8 @@ class Page
     /**
      * Helper method to return a page.
      *
-     * @param  string $url the url of the page
-     * @param bool    $all
+     * @param string $url the url of the page
+     * @param bool   $all
      *
      * @return \Grav\Common\Page\Page page you were looking for if it exists
      */
@@ -1829,6 +2055,7 @@ class Page
     {
         /** @var Pages $pages */
         $pages = self::getGrav()['pages'];
+
         return $pages->dispatch($url, $all);
     }
 
@@ -1836,20 +2063,21 @@ class Page
      * Get a collection of pages in the current context.
      *
      * @param string|array $params
-     * @param boolean $pagination
+     * @param boolean      $pagination
+     *
      * @return Collection
      * @throws \InvalidArgumentException
      */
     public function collection($params = 'content', $pagination = true)
     {
         if (is_string($params)) {
-            $params = (array) $this->value('header.'.$params);
+            $params = (array)$this->value('header.' . $params);
         } elseif (!is_array($params)) {
             throw new \InvalidArgumentException('Argument should be either header variable name or array of parameters');
         }
 
         if (!isset($params['items'])) {
-            return array();
+            return [];
         }
 
         $collection = $this->evaluate($params['items']);
@@ -1866,7 +2094,7 @@ class Page
         $process_taxonomy = isset($params['url_taxonomy_filters']) ? $params['url_taxonomy_filters'] : $config->get('system.pages.url_taxonomy_filters');
 
         if ($process_taxonomy) {
-            foreach ((array) $config->get('site.taxonomies') as $taxonomy) {
+            foreach ((array)$config->get('site.taxonomies') as $taxonomy) {
                 if ($uri->param($taxonomy)) {
                     $items = explode(',', $uri->param($taxonomy));
                     $collection->setParams(['taxonomies' => [$taxonomy => $items]]);
@@ -1877,8 +2105,9 @@ class Page
                             continue;
                         }
                         foreach ($items as $item) {
-                            if (empty($page->taxonomy[$taxonomy])
-                                || !in_array(htmlspecialchars_decode($item, ENT_QUOTES), $page->taxonomy[$taxonomy])) {
+                            if (empty($page->taxonomy[$taxonomy]) || !in_array(htmlspecialchars_decode($item,
+                                    ENT_QUOTES), $page->taxonomy[$taxonomy])
+                            ) {
                                 $collection->remove();
                             }
                         }
@@ -1928,27 +2157,28 @@ class Page
      * @return mixed
      * @internal
      */
-    protected function evaluate($value)
+    public function evaluate($value)
     {
         // Parse command.
         if (is_string($value)) {
             // Format: @command.param
             $cmd = $value;
-            $params = array();
-        } elseif (is_array($value) && count($value) == 1) {
+            $params = [];
+        } elseif (is_array($value) && count($value) == 1 && !is_int(key($value))) {
             // Format: @command.param: { attr1: value1, attr2: value2 }
-            $cmd = (string) key($value);
-            $params = (array) current($value);
+            $cmd = (string)key($value);
+            $params = (array)current($value);
         } else {
             $result = [];
-            foreach($value as $key => $val) {
+            foreach ($value as $key => $val) {
                 if (is_int($key)) {
                     $result = $result + $this->evaluate($val)->toArray();
                 } else {
-                    $result = $result + $this->evaluate([$key=>$val])->toArray();
+                    $result = $result + $this->evaluate([$key => $val])->toArray();
                 }
 
             }
+
             return new Collection($result);
         }
 
@@ -1957,50 +2187,93 @@ class Page
             return $value;
         }
 
+        /** @var Pages $pages */
+        $pages = self::getGrav()['pages'];
+
         $parts = explode('.', $cmd);
         $current = array_shift($parts);
 
-        $results = null;
+        /** @var Collection $results */
+        $results = new Collection();
+
         switch ($current) {
             case '@self':
                 if (!empty($parts)) {
                     switch ($parts[0]) {
                         case 'modular':
+                            // @self.modular: false (alternative to @self.children)
                             if (!empty($params) && $params[0] === false) {
-                                $results = $this->children()->nonModular()->published();
+                                $results = $this->children()->nonModular();
                                 break;
                             }
-                            $results = $this->children()->modular()->published();
+                            $results = $this->children()->modular();
                             break;
                         case 'children':
-                            $results = $this->children()->nonModular()->published();
+                            $results = $this->children()->nonModular();
+                            break;
+                        case 'all':
+                            $results = $this->children();
+                            break;
+                        case 'parent':
+                            $collection = new Collection();
+                            $results = $collection->addPage($this->parent());
+                            break;
+                        case 'siblings':
+                            $results = $this->parent()->children()->remove($this->path());
+                            break;
+                        case 'descendants':
+                            $results = $pages->all($this)->remove($this->path())->nonModular();
                             break;
                     }
                 }
+
+                $results = $results->published();
                 break;
 
             case '@page':
+                $page = null;
+
                 if (!empty($params)) {
-                    /** @var Pages $pages */
-                    $pages = self::getGrav()['pages'];
+                    $page = $this->find($params[0]);
+                }
 
-                    list($what, $recurse) = array_pad($params, 2, null);
+                // safety check in case page is not found
+                if (!isset($page)) {
+                    return $results;
+                }
 
-                    if ($what == '@root') {
-                        $page = $pages->root();
-                    } else {
-                        $page = $this->find($what);
+                // Handle a @page.descendants
+                if (!empty($parts)) {
+                    switch ($parts[0]) {
+                        case 'self':
+                            $results = new Collection();
+                            $results = $results->addPage($page);
+                            break;
+
+                        case 'descendants':
+                            $results = $pages->all($page)->remove($page->path());
+                            break;
+
+                        case 'children':
+                            $results = $page->children();
+                            break;
                     }
+                } else {
+                    $results = $page->children();
+                }
 
-                    if ($page) {
-                        if ($recurse) {
-                            $results = $pages->all($page)->nonModular()->published();
-                        } else {
-                            $results = $page->children()->nonModular()->published();
-                        }
-                     }
+                $results = $results->nonModular()->published();
+
+                break;
+
+            case '@root':
+                if (!empty($parts) && $parts[0] == 'descendants') {
+                    $results = $pages->all($pages->root())->nonModular()->published();
+                } else {
+                    $results = $pages->root()->children()->nonModular()->published();
                 }
                 break;
+
 
             case '@taxonomy':
                 // Gets a collection of pages by using one of the following formats:
@@ -2031,6 +2304,7 @@ class Page
         if ($this->name) {
             return true;
         }
+
         return false;
     }
 
@@ -2052,6 +2326,7 @@ class Page
     public function exists()
     {
         $file = $this->file();
+
         return $file && $file->exists();
     }
 
@@ -2059,6 +2334,7 @@ class Page
      * Cleans the path.
      *
      * @param  string $path the path
+     *
      * @return string       the path
      */
     protected function cleanPath($path)
@@ -2067,6 +2343,7 @@ class Page
         if (strpos($lastchunk, ':') !== false) {
             $path = str_replace($lastchunk, '', $path);
         }
+
         return $path;
     }
 
@@ -2074,10 +2351,14 @@ class Page
      * Moves or copies the page in filesystem.
      *
      * @internal
+     *
+     * @param bool $reorder
+     *
+     * @throws Exception
      */
     protected function doRelocation($reorder)
     {
-        if (empty($this->_original) ) {
+        if (!$this->_original) {
             return;
         }
 
@@ -2089,7 +2370,7 @@ class Page
             $parent = $this->parent();
 
             // Extract visible children from the parent page.
-            $list = array();
+            $list = [];
             /** @var Page $page */
             foreach ($parent->children()->visible() as $page) {
                 if ($page->order()) {
@@ -2106,31 +2387,32 @@ class Page
 
             // Then add it back to the new location (if needed).
             if ($this->order()) {
-                array_splice($list, min($this->order()-1, count($list)), 0, array($this->path()));
+                array_splice($list, min($this->order() - 1, count($list)), 0, [$this->path()]);
             }
 
             // Reorder all moved pages.
             foreach ($list as $order => $path) {
                 if ($path == $this->path()) {
                     // Handle current page; we do want to change ordering number, but nothing else.
-                    $this->order($order+1);
+                    $this->order($order + 1);
                 } else {
                     // Handle all the other pages.
                     $page = $pages->get($path);
 
-                    if ($page && $page->exists() && $page->order() != $order+1) {
+                    if ($page && $page->exists() && !$page->_action && $page->order() != $order + 1) {
                         $page = $page->move($parent);
-                        $page->order($order+1);
+                        $page->order($order + 1);
                         $page->save(false);
                     }
                 }
             }
         }
-        if ($this->_action == 'move' && $this->_original->exists()) {
-            Folder::move($this->_original->path(), $this->path());
-        }
-        if ($this->_action == 'copy' && $this->_original->exists()) {
-            Folder::copy($this->_original->path(), $this->path());
+        if (is_dir($this->_original->path())) {
+            if ($this->_action == 'move') {
+                Folder::move($this->_original->path(), $this->path());
+            } elseif ($this->_action == 'copy') {
+                Folder::copy($this->_original->path(), $this->path());
+            }
         }
 
         if ($this->name() != $this->_original->name()) {
@@ -2140,7 +2422,6 @@ class Page
             }
         }
 
-        $this->_action = null;
         $this->_original = null;
     }
 
@@ -2158,7 +2439,7 @@ class Page
                 }
             }
             // publish if required, if not clear cache right before page is published
-            if ($this->publishDate() != $this->modified() && $this->publishDate() > time()) {
+            if ($this->publishDate() && $this->publishDate() && $this->publishDate() > time()) {
                 $this->published(false);
                 self::getGrav()['cache']->setLifeTime($this->publishDate());
             }
