@@ -3,17 +3,17 @@ namespace Grav\Common\Twig;
 
 use Grav\Common\Grav;
 use Grav\Common\Config\Config;
+use Grav\Common\Language\Language;
 use Grav\Common\Page\Page;
-use Grav\Common\Inflector;
-use Grav\Common\Utils;
 use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
+use RocketTheme\Toolbox\Event\Event;
 
 /**
  * The Twig object handles all the Twig template rendering for Grav. It's a singleton object
  * that is optimized so that it only needs to be initialized once and can be reused for individual
  * page template rendering as well as the main site template rendering.
  *
- * @author RocketTheme
+ * @author  RocketTheme
  * @license MIT
  */
 class Twig
@@ -26,7 +26,7 @@ class Twig
     /**
      * @var array
      */
-    public $twig_vars;
+    public $twig_vars = [];
 
     /**
      * @var array
@@ -56,6 +56,8 @@ class Twig
 
     /**
      * Constructor
+     *
+     * @param Grav $grav
      */
     public function __construct(Grav $grav)
     {
@@ -74,18 +76,20 @@ class Twig
             $config = $this->grav['config'];
             /** @var UniformResourceLocator $locator */
             $locator = $this->grav['locator'];
-            $debugger = $this->grav['debugger'];
 
             /** @var Language $language */
             $language = $this->grav['language'];
 
             $active_language = $language->getActive();
 
-            $language_append = $active_language ? '/'.$active_language : '';
+            $language_append = '';
+            if ($language->getDefault() != $active_language || $config->get('system.languages.include_default_lang') === true) {
+                $language_append = $active_language ? '/' . $active_language : '';
+            }
 
             // handle language templates if available
             if ($language->enabled()) {
-                $lang_templates = $locator->findResource('theme://templates/'.($active_language ? $active_language : $language->getDefault()));
+                $lang_templates = $locator->findResource('theme://templates/' . ($active_language ? $active_language : $language->getDefault()));
                 if ($lang_templates) {
                     $this->twig_paths[] = $lang_templates;
                 }
@@ -96,8 +100,8 @@ class Twig
             $this->grav->fireEvent('onTwigTemplatePaths');
 
             $this->loader = new \Twig_Loader_Filesystem($this->twig_paths);
-            $this->loaderArray = new \Twig_Loader_Array(array());
-            $loader_chain = new \Twig_Loader_Chain(array($this->loaderArray, $this->loader));
+            $this->loaderArray = new \Twig_Loader_Array([]);
+            $loader_chain = new \Twig_Loader_Chain([$this->loaderArray, $this->loader]);
 
             $params = $config->get('system.twig');
             if (!empty($params['cache'])) {
@@ -112,7 +116,8 @@ class Twig
                         return new \Twig_Function_Function($name);
                     }
 
-                    return new \Twig_Function_Function(function() {});
+                    return new \Twig_Function_Function(function () {
+                    });
                 });
             }
 
@@ -122,7 +127,8 @@ class Twig
                         return new \Twig_Filter_Function($name);
                     }
 
-                    return new \Twig_Filter_Function(function() {});
+                    return new \Twig_Filter_Function(function () {
+                    });
                 });
             }
 
@@ -141,21 +147,21 @@ class Twig
             $this->grav->fireEvent('onTwigExtensions');
 
             // Set some standard variables for twig
-            $this->twig_vars = array(
-                'config' => $config,
-                'uri' => $this->grav['uri'],
-                'base_dir' => rtrim(ROOT_DIR, '/'),
-                'base_url' => $this->grav['base_url'] . $language_append,
-                'base_url_simple' => $this->grav['base_url'],
-                'base_url_absolute' => $this->grav['base_url_absolute'] . $language_append,
-                'base_url_relative' => $this->grav['base_url_relative'] . $language_append,
-                'theme_dir' => $locator->findResource('theme://'),
-                'theme_url' => $this->grav['base_url'] .'/'. $locator->findResource('theme://', false),
-                'site' => $config->get('site'),
-                'assets' => $this->grav['assets'],
-                'taxonomy' => $this->grav['taxonomy'],
-                'browser' => $this->grav['browser'],
-            );
+            $this->twig_vars = $this->twig_vars + [
+                    'config'            => $config,
+                    'uri'               => $this->grav['uri'],
+                    'base_dir'          => rtrim(ROOT_DIR, '/'),
+                    'base_url'          => $this->grav['base_url'] . $language_append,
+                    'base_url_simple'   => $this->grav['base_url'],
+                    'base_url_absolute' => $this->grav['base_url_absolute'] . $language_append,
+                    'base_url_relative' => $this->grav['base_url_relative'] . $language_append,
+                    'theme_dir'         => $locator->findResource('theme://'),
+                    'theme_url'         => $this->grav['base_url'] . '/' . $locator->findResource('theme://', false),
+                    'site'              => $config->get('site'),
+                    'assets'            => $this->grav['assets'],
+                    'taxonomy'          => $this->grav['taxonomy'],
+                    'browser'           => $this->grav['browser'],
+                ];
         }
     }
 
@@ -193,6 +199,7 @@ class Twig
      *
      * @param  Page   $item    The page item to render
      * @param  string $content Optional content override
+     *
      * @return string          The rendered output
      * @throws \Twig_Error_Loader
      */
@@ -201,7 +208,7 @@ class Twig
         $content = $content !== null ? $content : $item->content();
 
         // override the twig header vars for local resolution
-        $this->grav->fireEvent('onTwigPageVariables');
+        $this->grav->fireEvent('onTwigPageVariables', new Event(['page' => $item]));
         $twig_vars = $this->twig_vars;
 
         $twig_vars['page'] = $item;
@@ -211,6 +218,8 @@ class Twig
         $local_twig = clone($this->twig);
         $modular_twig = $item->modularTwig();
         $process_twig = isset($item->header()->process['twig']) ? $item->header()->process['twig'] : false;
+
+        $output = '';
 
         try {
             // Process Modular Twig
@@ -238,10 +247,11 @@ class Twig
      * and optional array of variables
      *
      * @param string $template template to render with
-     * @param array $vars      Optional variables
+     * @param array  $vars     Optional variables
+     *
      * @return string
      */
-    public function processTemplate($template, $vars = array())
+    public function processTemplate($template, $vars = [])
     {
         // override the twig header vars for local resolution
         $this->grav->fireEvent('onTwigTemplateVariables');
@@ -262,11 +272,12 @@ class Twig
      * Process a Twig template directly by using a Twig string
      * and optional array of variables
      *
-     * @param string $string  string to render.
-     * @param array $vars     Optional variables
+     * @param string $string string to render.
+     * @param array  $vars   Optional variables
+     *
      * @return string
      */
-    public function processString($string, array $vars = array())
+    public function processString($string, array $vars = [])
     {
         // override the twig header vars for local resolution
         $this->grav->fireEvent('onTwigStringVariables');
@@ -289,6 +300,7 @@ class Twig
      * page and handles all the layout for the site display.
      *
      * @param string $format Output format (defaults to HTML).
+     *
      * @return string the rendered output
      * @throws \RuntimeException
      */
@@ -299,7 +311,6 @@ class Twig
         $pages = $this->grav['pages'];
         $page = $this->grav['page'];
         $content = $page->content();
-        $config = $this->grav['config'];
 
         $twig_vars = $this->twig_vars;
 
@@ -324,9 +335,9 @@ class Twig
         } catch (\Twig_Error_Loader $e) {
             $error_msg = $e->getMessage();
             // Try html version of this template if initial template was NOT html
-            if ($ext != '.html'.TWIG_EXT) {
+            if ($ext != '.html' . TWIG_EXT) {
                 try {
-                    $output = $this->twig->render($page->template().'.html'.TWIG_EXT, $twig_vars);
+                    $output = $this->twig->render($page->template() . '.html' . TWIG_EXT, $twig_vars);
                 } catch (\Twig_Error_Loader $e) {
                     throw new \RuntimeException($error_msg, 400, $e);
                 }
@@ -343,6 +354,7 @@ class Twig
      * the one being passed in
      *
      * @param  string $template the template name
+     *
      * @return string           the template name
      */
     public function template($template)
