@@ -1,5 +1,6 @@
 <?php
 namespace Grav\Common\Data;
+
 use Grav\Common\GravTrait;
 use Symfony\Component\Yaml\Exception\ParseException;
 use Symfony\Component\Yaml\Parser;
@@ -30,14 +31,19 @@ class Validation
             return;
         }
 
+        // special case for files, value is never empty and errors with code 4 instead
+        if (empty($validate['required']) && $field['type'] == 'file' && (isset($value['error']) && ($value['error'] == UPLOAD_ERR_NO_FILE || in_array(UPLOAD_ERR_NO_FILE, $value['error'])))) {
+            return;
+        }
+
         // Get language class
         $language = self::getGrav()['language'];
 
         // Validate type with fallback type text.
         $type = (string) isset($field['validate']['type']) ? $field['validate']['type'] : $field['type'];
         $method = 'type'.strtr($type, '-', '_');
-        $name = ucfirst($field['label'] ? $field['label'] : $field['name']);
-        $message = (string) isset($field['validate']['message']) ? $field['validate']['message'] : 'Invalid input in "' . $language->translate($name) . '""';
+        $name = ucfirst(isset($field['label']) ? $field['label'] : $field['name']);
+        $message = (string) isset($field['validate']['message']) ? $language->translate($field['validate']['message']) : $language->translate('FORM.INVALID_INPUT', null, true) . ' "' . $language->translate($name) . '"';
 
         if (method_exists(__CLASS__, $method)) {
             $success = self::$method($value, $validate, $field);
@@ -74,6 +80,11 @@ class Validation
 
         // If value isn't required, we will return null if empty value is given.
         if (empty($validate['required']) && ($value === null || $value === '')) {
+            return null;
+        }
+
+        // special case for files, value is never empty and errors with code 4 instead
+        if (empty($validate['required']) && $field['type'] == 'file' && (isset($value['error']) && ($value['error'] == UPLOAD_ERR_NO_FILE || in_array(UPLOAD_ERR_NO_FILE, $value['error'])))) {
             return null;
         }
 
@@ -255,6 +266,32 @@ class Validation
     public static function typeToggle($value, array $params, array $field)
     {
         return self::typeArray((array) $value, $params, $field);
+    }
+
+    /**
+     * Custom input: file
+     *
+     * @param  mixed  $value   Value to be validated.
+     * @param  array  $params  Validation parameters.
+     * @param  array  $field   Blueprint for the field.
+     * @return bool   True if validation succeeded.
+     */
+    public static function typeFile($value, array $params, array $field)
+    {
+        return self::typeArray((array) $value, $params, $field);
+    }
+
+    protected static function filterFile($value, array $params, array $field)
+    {
+        if (isset($field['multiple']) && $field['multiple'] === true) {
+            return (array) $value;
+        }
+
+        if (is_array($value)) {
+            return reset($value);
+        }
+
+        return $value;
     }
 
     /**
@@ -522,6 +559,10 @@ class Validation
         $options = isset($field['options']) ? array_keys($field['options']) : array();
         $multi = isset($field['multiple']) ? $field['multiple'] : false;
 
+        if (count($values) == 1 && isset($values[0]) && $values[0] == '') {
+            return null;
+        }
+
         if ($options) {
             $useKey = isset($field['use']) && $field['use'] == 'keys';
             foreach ($values as $key => $value) {
@@ -588,6 +629,10 @@ class Validation
 
     public static function validateRequired($value, $params)
     {
+        if (is_string($value)) {
+            $value = trim($value);
+        }
+
         return (bool) $params !== true || !empty($value);
     }
 
