@@ -323,13 +323,24 @@ class GPM extends Iterator
             return $found;
         }
 
-        foreach ($this->getRepositoryThemes() as $slug => $theme) {
+        $themes = $this->getRepositoryThemes();
+        $plugins = $this->getRepositoryPlugins();
+
+        if (!$themes && !$plugins) {
+            if (!is_writable(ROOT_DIR . '/cache/gpm')) {
+                throw new \RuntimeException("The cache/gpm folder is not writable. Please check the folder permissions.");
+            }
+
+            throw new \RuntimeException("GPM not reachable. Please check your internet connection or check the Grav site is reachable");
+        }
+
+        if ($themes) foreach ($themes as $slug => $theme) {
             if ($search == $slug || $search == $theme->name) {
                 return $theme;
             }
         }
 
-        foreach ($this->getRepositoryPlugins() as $slug => $plugin) {
+        if ($plugins) foreach ($plugins as $slug => $plugin) {
             if ($search == $slug || $search == $plugin->name) {
                 return $plugin;
             }
