@@ -3,10 +3,9 @@
 namespace Grav\Common\GPM\Local;
 
 use Grav\Common\GPM\Common\AbstractPackageCollection as BaseCollection;
-use Grav\Common\GPM\Local\Package;
 
-abstract class AbstractPackageCollection extends BaseCollection {
-
+abstract class AbstractPackageCollection extends BaseCollection
+{
     public function __construct($items)
     {
         foreach ($items as $name => $data) {
