@@ -12,10 +12,12 @@ class Themes extends AbstractPackageCollection
      */
     protected $type = 'themes';
 
-    protected $repository = 'http://getgrav.org/downloads/themes.json';
+    protected $repository = 'https://getgrav.org/downloads/themes.json';
 
     /**
      * Local Themes Constructor
+     * @param bool $refresh
+     * @param callable $callback Either a function or callback in array notation
      */
     public function __construct($refresh = false, $callback = null)
     {
