@@ -1,6 +1,7 @@
 <?php
 namespace Grav\Common\Page\Medium;
 
+use Grav\Common\Utils;
 use Grav\Common\Data\Blueprint;
 
 class ImageMedium extends Medium
@@ -41,7 +42,8 @@ class ImageMedium extends Medium
     public static $magic_actions = [
         'resize', 'forceResize', 'cropResize', 'crop', 'zoomCrop',
         'negate', 'brightness', 'contrast', 'grayscale', 'emboss',
-        'smooth', 'sharp', 'edge', 'colorize', 'sepia', 'enableProgressive'
+        'smooth', 'sharp', 'edge', 'colorize', 'sepia', 'enableProgressive',
+        'rotate', 'flip', 'fixOrientation'
     ];
 
     /**
@@ -52,7 +54,6 @@ class ImageMedium extends Medium
         'forceResize' => [ 0, 1 ],
         'cropResize' => [ 0, 1 ],
         'crop' => [ 0, 1, 2, 3 ],
-        'cropResize' => [ 0, 1 ],
         'zoomCrop' => [ 0, 1 ]
     ];
 
@@ -136,7 +137,15 @@ class ImageMedium extends Medium
      */
     public function url($reset = true)
     {
-        $output = preg_replace('|^' . GRAV_ROOT . '|', '', $this->saveImage());
+        $image_path = self::$grav['locator']->findResource('cache://images', true);
+        $image_dir = self::$grav['locator']->findResource('cache://images', false);
+        $saved_image_path = $this->saveImage();
+
+        $output = preg_replace('|^' . preg_quote(GRAV_ROOT) . '|', '', $saved_image_path);
+
+        if (Utils::startsWith($output, $image_path)) {
+            $output = '/' . $image_dir . preg_replace('|^' . preg_quote($image_path) . '|', '', $output);
+        }
 
         if ($reset) {
             $this->reset();
@@ -155,6 +164,7 @@ class ImageMedium extends Medium
         if (!$this->image) {
             $this->image();
         }
+
         return $this;
     }
 
@@ -260,6 +270,8 @@ class ImageMedium extends Medium
 
         if ($this->image) {
             $this->image();
+            $this->image->clearOperations(); // Clear previously applied operations
+            $this->querystring('');
             $this->filter();
         }
 
@@ -311,19 +323,23 @@ class ImageMedium extends Medium
     }
 
     /**
-     * Sets the quality of the image
+     * Sets or gets the quality of the image
      *
      * @param  int $quality 0-100 quality
      * @return Medium
      */
-    public function quality($quality)
+    public function quality($quality = null)
     {
-        if (!$this->image) {
-            $this->image();
+        if ($quality) {
+            if (!$this->image) {
+                $this->image();
+            }
+
+            $this->quality = $quality;
+            return $this;
         }
 
-        $this->quality = $quality;
-        return $this;
+        return $this->quality;
     }
 
     /**
@@ -357,6 +373,48 @@ class ImageMedium extends Medium
         }
 
         return empty($this->sizes) ? '100vw' : $this->sizes;
+    }
+
+    /**
+     * Allows to set the width attribute from Markdown or Twig
+     * Examples: ![Example](myimg.png?width=200&height=400)
+     *           ![Example](myimg.png?resize=100,200&width=100&height=200)
+     *           ![Example](myimg.png?width=auto&height=auto)
+     *           ![Example](myimg.png?width&height)
+     *           {{ page.media['myimg.png'].width().height().html }}
+     *           {{ page.media['myimg.png'].resize(100,200).width(100).height(200).html }}
+     *
+     * @param mixed $value A value or 'auto' or empty to use the width of the image
+     * @return $this
+     */
+    public function width($value = 'auto')
+    {
+        if (!$value || $value == 'auto')
+            $this->attributes['width'] = $this->get('width');
+        else
+            $this->attributes['width'] = $value;
+        return $this;
+    }
+
+    /**
+     * Allows to set the height attribute from Markdown or Twig
+     * Examples: ![Example](myimg.png?width=200&height=400)
+     *           ![Example](myimg.png?resize=100,200&width=100&height=200)
+     *           ![Example](myimg.png?width=auto&height=auto)
+     *           ![Example](myimg.png?width&height)
+     *           {{ page.media['myimg.png'].width().height().html }}
+     *           {{ page.media['myimg.png'].resize(100,200).width(100).height(200).html }}
+     *
+     * @param mixed $value A value or 'auto' or empty to use the height of the image
+     * @return $this
+     */
+    public function height($value = 'auto')
+    {
+        if (!$value || $value == 'auto')
+            $this->attributes['height'] = $this->get('height');
+        else
+            $this->attributes['height'] = $value;
+        return $this;
     }
 
     /**
@@ -470,4 +528,28 @@ class ImageMedium extends Medium
             $this->__call($method, $params);
         }
     }
+
+    /**
+     * Return the image higher quality version
+     *
+     * @return ImageMedium the alternative version with higher quality
+     */
+    public function higherQualityAlternative()
+    {
+        if ($this->alternatives) {
+            $max = reset($this->alternatives);
+            foreach($this->alternatives as $alternative)
+            {
+                if($alternative->quality() > $max->quality())
+                {
+                    $max = $alternative;
+                }
+            }
+
+            return $max;
+        } else {
+            return $this;
+        }
+    }
+
 }
