@@ -118,7 +118,7 @@ class Medium extends Data implements RenderableInterface
      * Return PATH to file.
      *
      * @param bool $reset
-     * @return  string path to file
+     * @return string path to file
      */
     public function path($reset = true)
     {
@@ -137,7 +137,7 @@ class Medium extends Data implements RenderableInterface
      */
     public function url($reset = true)
     {
-        $output = preg_replace('|^' . GRAV_ROOT . '|', '', $this->get('filepath'));
+        $output = preg_replace('|^' . preg_quote(GRAV_ROOT) . '|', '', $this->get('filepath'));
 
         if ($reset) {
             $this->reset();
@@ -155,7 +155,7 @@ class Medium extends Data implements RenderableInterface
      */
     public function querystring($querystring = null, $withQuestionmark = true)
     {
-        if ($querystring) {
+        if (!is_null($querystring)) {
             $this->set('querystring', ltrim($querystring, '?&'));
 
             foreach ($this->alternatives as $alt) {
@@ -209,9 +209,14 @@ class Medium extends Data implements RenderableInterface
 
         $style = '';
         foreach ($this->styleAttributes as $key => $value) {
-            $style .= $key . ': ' . $value . ';';
+            if (is_numeric($key)) // Special case for inline style attributes, refer to style() method
+                $style .= $value;
+            else
+                $style .= $key . ': ' . $value . ';';
         }
-        $attributes['style'] = $style;
+        if ($style) {
+            $attributes['style'] = $style;
+        }
 
         if (empty($attributes['title'])) {
             if (!empty($title)) {
@@ -389,6 +394,51 @@ class Medium extends Data implements RenderableInterface
     }
 
     /**
+     * Add a class to the element from Markdown or Twig
+     * Example: ![Example](myimg.png?classes=float-left) or ![Example](myimg.png?classes=myclass1,myclass2)
+     *
+     * @return $this
+     */
+    public function classes()
+    {
+        $classes = func_get_args();
+        if (!empty($classes)) {
+            $this->attributes['class'] = implode(',', (array)$classes);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Add an id to the element from Markdown or Twig
+     * Example: ![Example](myimg.png?id=primary-img)
+     *
+     * @param $id
+     * @return $this
+     */
+    public function id($id)
+    {
+        if (is_string($id)) {
+            $this->attributes['id'] = trim($id);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Allows to add an inline style attribute from Markdown or Twig
+     * Example: ![Example](myimg.png?style=float:left)
+     *
+     * @param string $style
+     * @return $this
+     */
+    public function style($style)
+    {
+        $this->styleAttributes[] = rtrim($style, ';') . ';';
+        return $this;
+    }
+
+    /**
      * Allow any action to be called on this medium from twig or markdown
      *
      * @param string $method
@@ -440,4 +490,5 @@ class Medium extends Data implements RenderableInterface
 
         return $this->_thumbnail;
     }
+
 }
