@@ -1,10 +1,15 @@
 <?php
 namespace Grav\Common\Service;
 
+use Grav\Common\Config\CompiledBlueprints;
+use Grav\Common\Config\CompiledConfig;
+use Grav\Common\Config\CompiledLanguages;
 use Grav\Common\Config\Config;
+use Grav\Common\Config\ConfigFileFinder;
+use Grav\Common\Config\Setup;
 use Pimple\Container;
 use Pimple\ServiceProviderInterface;
-use RocketTheme\Toolbox\Blueprints\Blueprints;
+use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 
 /**
  * The Config class contains configuration information.
@@ -14,51 +19,134 @@ use RocketTheme\Toolbox\Blueprints\Blueprints;
  */
 class ConfigServiceProvider implements ServiceProviderInterface
 {
-    private $environment;
-    private $setup;
-
     public function register(Container $container)
     {
-        $self = $this;
-
-        // Pre-load setup.php as it contains our initial configuration.
-        $file = GRAV_ROOT . '/setup.php';
-        $this->setup = is_file($file) ? (array) include $file : [];
-        $this->environment = isset($this->setup['environment']) ? $this->setup['environment'] : null;
-
-        $container['blueprints'] = function ($c) use ($self) {
-            return $self->loadMasterBlueprints($c);
+        $container['setup'] = function ($c) {
+            return static::setup($c);
         };
 
-        $container['config'] = function ($c) use ($self) {
-            return $self->loadMasterConfig($c);
+        $container['blueprints'] = function ($c) {
+            return static::blueprints($c);
+        };
+
+        $container['config'] = function ($c) {
+            return static::load($c);
+        };
+
+        $container['languages'] = function ($c) {
+            return static::languages($c);
         };
     }
 
-    public function loadMasterConfig(Container $container)
+    public static function setup(Container $container)
     {
-        $environment = $this->getEnvironment($container);
-
-        $config = new Config($this->setup, $container, $environment);
-
-        return $config;
+        return new Setup($container);
     }
 
-    public function loadMasterBlueprints(Container $container)
+    public static function blueprints(Container $container)
     {
-        $environment = $this->getEnvironment($container);
-        $file = CACHE_DIR . 'compiled/blueprints/master-'.$environment.'.php';
-        $data = is_file($file) ? (array) include $file : [];
+        /** Setup $setup */
+        $setup = $container['setup'];
 
-        return new Blueprints($data, $container);
+        /** @var UniformResourceLocator $locator */
+        $locator = $container['locator'];
+
+        $cache =  $locator->findResource('cache://compiled/blueprints', true, true);
+
+        $files = [];
+        $paths = $locator->findResources('blueprints://config');
+        $files += (new ConfigFileFinder)->locateFiles($paths);
+        $paths = $locator->findResources('plugins://');
+        $files += (new ConfigFileFinder)->setBase('plugins')->locateInFolders($paths, 'blueprints');
+
+        $blueprints = new CompiledBlueprints($cache, $files, GRAV_ROOT);
+
+        return $blueprints->name("master-{$setup->environment}")->load();
     }
 
-    public function getEnvironment(Container $container)
+    public static function load(Container $container)
     {
-        if (!isset($this->environment)) {
-            $this->environment = $container['uri']->environment();
+        /** Setup $setup */
+        $setup = $container['setup'];
+
+        /** @var UniformResourceLocator $locator */
+        $locator = $container['locator'];
+
+        $cache =  $locator->findResource('cache://compiled/config', true, true);
+
+        $files = [];
+        $paths = $locator->findResources('config://');
+        $files += (new ConfigFileFinder)->locateFiles($paths);
+        $paths = $locator->findResources('plugins://');
+        $files += (new ConfigFileFinder)->setBase('plugins')->locateInFolders($paths);
+
+        $config = new CompiledConfig($cache, $files, GRAV_ROOT);
+        $config->setBlueprints(function() use ($container) {
+            return $container['blueprints'];
+        });
+
+        return $config->name("master-{$setup->environment}")->load();
+    }
+
+    public static function languages(Container $container)
+    {
+        /** Setup $setup */
+        $setup = $container['setup'];
+
+        /** @var Config $config */
+        $config = $container['config'];
+
+        /** @var UniformResourceLocator $locator */
+        $locator = $container['locator'];
+
+        $cache = $locator->findResource('cache://compiled/languages', true, true);
+        $files = [];
+
+        // Process languages only if enabled in configuration.
+        if ($config->get('system.languages.translations', true)) {
+            $paths = $locator->findResources('languages://');
+            $files += (new ConfigFileFinder)->locateFiles($paths);
+            $paths = $locator->findResources('plugins://');
+            $files += (new ConfigFileFinder)->setBase('plugins')->locateInFolders($paths, 'languages');
+            $paths = static::pluginFolderPaths($paths, 'languages');
+            $files += (new ConfigFileFinder)->locateFiles($paths);
         }
 
-        return $this->environment;
+        $languages = new CompiledLanguages($cache, $files, GRAV_ROOT);
+
+        return $languages->name("master-{$setup->environment}")->load();
     }
+
+    /**
+     * Find specific paths in plugins
+     *
+     * @param $plugins
+     * @param $folder_path
+     * @return array
+     */
+    private static function pluginFolderPaths($plugins, $folder_path)
+    {
+        $paths = [];
+
+        foreach ($plugins as $path) {
+            $iterator = new \DirectoryIterator($path);
+
+            /** @var \DirectoryIterator $directory */
+            foreach ($iterator as $directory) {
+                if (!$directory->isDir() || $directory->isDot()) {
+                    continue;
+                }
+
+                // Path to the languages folder
+                $lang_path = $directory->getPathName() . '/' . $folder_path;
+
+                // If this folder exists, add it to the list of paths
+                if (file_exists($lang_path)) {
+                    $paths []= $lang_path;
+                }
+            }
+        }
+        return $paths;
+    }
+
 }
