@@ -1,7 +1,7 @@
 <?php
 namespace Grav\Common\Service;
 
-use Grav\Common\Config\Config;
+use Grav\Common\Config\Setup;
 use Pimple\Container;
 use RocketTheme\Toolbox\DI\ServiceProviderInterface;
 use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
@@ -14,18 +14,18 @@ class StreamsServiceProvider implements ServiceProviderInterface
     public function register(Container $container)
     {
         $container['locator'] = function($c) {
-            $locator = new UniformResourceLocator(ROOT_DIR);
+            $locator = new UniformResourceLocator(GRAV_ROOT);
 
-            /** @var Config $config */
-            $config = $c['config'];
-            $config->initializeLocator($locator);
+            /** @var Setup $setup */
+            $setup = $c['setup'];
+            $setup->initializeLocator($locator);
 
             return $locator;
         };
 
         $container['streams'] = function($c) {
-            /** @var Config $config */
-            $config = $c['config'];
+            /** @var Setup $setup */
+            $setup = $c['setup'];
 
             /** @var UniformResourceLocator $locator */
             $locator = $c['locator'];
@@ -34,7 +34,7 @@ class StreamsServiceProvider implements ServiceProviderInterface
             Stream::setLocator($locator);
             ReadOnlyStream::setLocator($locator);
 
-            return new StreamBuilder($config->getStreams($c));
+            return new StreamBuilder($setup->getStreams($c));
         };
     }
 }
