@@ -10,10 +10,10 @@ use Grav\Common\Utils;
 /**
  * User object
  *
- * @property mixed authenticated
- * @property mixed password
+ * @property mixed       authenticated
+ * @property mixed       password
  * @property bool|string hashed_password
- * @author RocketTheme
+ * @author  RocketTheme
  * @license MIT
  */
 class User extends Data
@@ -26,11 +26,15 @@ class User extends Data
      * Always creates user object. To check if user exists, use $this->exists().
      *
      * @param string $username
+     *
      * @return User
      */
     public static function load($username)
     {
         $locator = self::getGrav()['locator'];
+
+        // force lowercase of username
+        $username = strtolower($username);
 
         $blueprints = new Blueprints('blueprints://');
         $blueprint = $blueprints->get('user/account');
@@ -39,6 +43,9 @@ class User extends Data
         $content = $file->content();
         if (!isset($content['username'])) {
             $content['username'] = $username;
+        }
+        if (!isset($content['state'])) {
+            $content['state'] = 'enabled';
         }
         $user = new User($content, $blueprint);
         $user->file($file);
@@ -50,7 +57,8 @@ class User extends Data
      * Remove user account.
      *
      * @param string $username
-     * @return bool True is the action was performed
+     *
+     * @return bool True if the action was performed
      */
     public static function remove($username)
     {
@@ -67,7 +75,8 @@ class User extends Data
      *
      * If user password needs to be updated, new information will be saved.
      *
-     * @param string $password  Plaintext password.
+     * @param string $password Plaintext password.
+     *
      * @return bool
      */
     public function authenticate($password)
@@ -80,7 +89,11 @@ class User extends Data
                 // Plain-text passwords do not match, we know we should fail but execute
                 // verify to protect us from timing attacks and return false regardless of
                 // the result
-                Authentication::verify($password, self::getGrav()['config']->get('system.security.default_hash'));
+                Authentication::verify(
+                    $password,
+                    self::getGrav()['config']->get('system.security.default_hash')
+                );
+
                 return false;
             } else {
                 // Plain-text does match, we can update the hash and proceed
@@ -104,7 +117,7 @@ class User extends Data
             $this->save();
         }
 
-        return (bool) $result;
+        return (bool)$result;
     }
 
     /**
@@ -130,7 +143,8 @@ class User extends Data
     /**
      * Checks user authorization to the action.
      *
-     * @param  string  $action
+     * @param  string $action
+     *
      * @return bool
      */
     public function authorize($action)
@@ -139,7 +153,33 @@ class User extends Data
             return false;
         }
 
-        return Utils::isPositive($this->get("access.{$action}"));
+        if (isset($this->state) && $this->state !== 'enabled') {
+            return false;
+        }
+
+        $return = false;
+
+        //Check group access level
+        $groups = $this->get('groups');
+        if ($groups) {
+            foreach ((array)$groups as $group) {
+                $permission = self::getGrav()['config']->get("groups.{$group}.access.{$action}");
+                $return = Utils::isPositive($permission);
+                if ($return === true) {
+                    break;
+                }
+            }
+        }
+
+        //Check user access level
+        if ($this->get('access')) {
+            if (Utils::resolve($this->get('access'), $action) !== null) {
+                $permission = $this->get("access.{$action}");
+                $return = Utils::isPositive($permission);
+            }
+        }
+
+        return $return;
     }
 
     /**
@@ -147,6 +187,7 @@ class User extends Data
      * Ensures backwards compatibility
      *
      * @param  string $action
+     *
      * @deprecated use authorize()
      * @return bool
      */
