@@ -2,11 +2,11 @@
 namespace Grav\Console\Cli;
 
 use Grav\Common\Filesystem\Folder;
-use Grav\Console\ConsoleTrait;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Formatter\OutputFormatterStyle;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Formatter\OutputFormatterStyle;
 
 /**
  * Class CleanCommand
@@ -14,12 +14,18 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 class CleanCommand extends Command
 {
-    use ConsoleTrait;
+    /* @var InputInterface $output */
+    protected $input;
+
+    /* @var OutputInterface $output */
+    protected $output;
 
     /**
      * @var array
      */
     protected $paths_to_remove = [
+        'codeception.yml',
+        'tests/',
         'user/plugins/email/vendor/swiftmailer/swiftmailer/.travis.yml',
         'user/plugins/email/vendor/swiftmailer/swiftmailer/build.xml',
         'user/plugins/email/vendor/swiftmailer/swiftmailer/composer.json',
@@ -89,7 +95,7 @@ class CleanCommand extends Command
         'vendor/maximebf/debugbar/bower.json',
         'vendor/maximebf/debugbar/composer.json',
         'vendor/maximebf/debugbar/.bowerrc',
-        'vendor/maximebf/debugbar/src/Debugbar/Resources/vendor',
+        'vendor/maximebf/debugbar/src/DebugBar/Resources/vendor',
         'vendor/maximebf/debugbar/demo',
         'vendor/maximebf/debugbar/docs',
         'vendor/maximebf/debugbar/tests',
@@ -127,28 +133,34 @@ class CleanCommand extends Command
         'vendor/rockettheme/toolbox/.travis.yml',
         'vendor/rockettheme/toolbox/composer.json',
         'vendor/rockettheme/toolbox/phpunit.xml',
-        'vendor/symfony/console/Symfony/Component/Console/composer.json',
-        'vendor/symfony/console/Symfony/Component/Console/phpunit.xml.dist',
-        'vendor/symfony/console/Symfony/Component/Console/.gitignore',
-        'vendor/symfony/console/Symfony/Component/Console/.git',
-        'vendor/symfony/console/Symfony/Component/Console/Tests',
-        'vendor/symfony/event-dispatcher/Symfony/Component/EventDispatcher/.git',
-        'vendor/symfony/event-dispatcher/Symfony/Component/EventDispatcher/.gitignore',
-        'vendor/symfony/event-dispatcher/Symfony/Component/EventDispatcher/composer.json',
-        'vendor/symfony/event-dispatcher/Symfony/Component/EventDispatcher/phpunit.xml.dist',
-        'vendor/symfony/event-dispatcher/Symfony/Component/EventDispatcher/Tests',
-        'vendor/symfony/yaml/Symfony/Component/Yaml/composer.json',
-        'vendor/symfony/yaml/Symfony/Component/Yaml/phpunit.xml.dist',
-        'vendor/symfony/yaml/Symfony/Component/Yaml/.gitignore',
-        'vendor/symfony/yaml/Symfony/Component/Yaml/.git',
-        'vendor/symfony/yaml/Symfony/Component/Yaml/Tests',
-        'vendor/tracy/tracy/.gitattributes',
-        'vendor/tracy/tracy/.travis.yml',
-        'vendor/tracy/tracy/composer.json',
-        'vendor/tracy/tracy/.gitignore',
-        'vendor/tracy/tracy/.git',
-        'vendor/tracy/tracy/examples',
-        'vendor/tracy/tracy/tests',
+        'vendor/symfony/console/composer.json',
+        'vendor/symfony/console/phpunit.xml.dist',
+        'vendor/symfony/console/.gitignore',
+        'vendor/symfony/console/.git',
+        'vendor/symfony/console/Tester',
+        'vendor/symfony/console/Tests',
+        'vendor/symfony/event-dispatcher/.git',
+        'vendor/symfony/event-dispatcher/.gitignore',
+        'vendor/symfony/event-dispatcher/composer.json',
+        'vendor/symfony/event-dispatcher/phpunit.xml.dist',
+        'vendor/symfony/event-dispatcher/Tests',
+        'vendor/symfony/polyfill-iconv/.git',
+        'vendor/symfony/polyfill-iconv/.gitignore',
+        'vendor/symfony/polyfill-iconv/composer.json',
+        'vendor/symfony/polyfill-mbstring/.git',
+        'vendor/symfony/polyfill-mbstring/.gitignore',
+        'vendor/symfony/polyfill-mbstring/composer.json',
+        'vendor/symfony/var-dumper/.git',
+        'vendor/symfony/var-dumper/.gitignore',
+        'vendor/symfony/var-dumper/composer.json',
+        'vendor/symfony/var-dumper/phpunit.xml.dist',
+        'vendor/symfony/var-dumper/Test',
+        'vendor/symfony/var-dumper/Tests',
+        'vendor/symfony/yaml/composer.json',
+        'vendor/symfony/yaml/phpunit.xml.dist',
+        'vendor/symfony/yaml/.gitignore',
+        'vendor/symfony/yaml/.git',
+        'vendor/symfony/yaml/Tests',
         'vendor/twig/twig/.editorconfig',
         'vendor/twig/twig/.travis.yml',
         'vendor/twig/twig/.gitignore',
@@ -172,14 +184,12 @@ class CleanCommand extends Command
     }
 
     /**
-     * @param InputInterface  $input
-     * @param OutputInterface $output
-     *
      * @return int|null|void
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->setupConsole($input, $output);
+
         $this->cleanPaths();
     }
 
@@ -187,12 +197,9 @@ class CleanCommand extends Command
     {
         $this->output->writeln('');
         $this->output->writeln('<red>DELETING</red>');
-
         $anything = false;
-
         foreach ($this->paths_to_remove as $path) {
             $path = ROOT_DIR . $path;
-
             if (is_dir($path) && @Folder::delete($path)) {
                 $anything = true;
                 $this->output->writeln('<red>dir:  </red>' . $path);
@@ -201,12 +208,30 @@ class CleanCommand extends Command
                 $this->output->writeln('<red>file: </red>' . $path);
             }
         }
-
         if (!$anything) {
             $this->output->writeln('');
             $this->output->writeln('<green>Nothing to clean...</green>');
         }
+    }
 
+        /**
+     * Set colors style definition for the formatter.
+     *
+     * @param InputInterface  $input
+     * @param OutputInterface $output
+     */
+    public function setupConsole(InputInterface $input, OutputInterface $output)
+    {
+        $this->input  = $input;
+        $this->output = $output;
+
+        $this->output->getFormatter()->setStyle('normal', new OutputFormatterStyle('white'));
+        $this->output->getFormatter()->setStyle('yellow', new OutputFormatterStyle('yellow', null, array('bold')));
+        $this->output->getFormatter()->setStyle('red', new OutputFormatterStyle('red', null, array('bold')));
+        $this->output->getFormatter()->setStyle('cyan', new OutputFormatterStyle('cyan', null, array('bold')));
+        $this->output->getFormatter()->setStyle('green', new OutputFormatterStyle('green', null, array('bold')));
+        $this->output->getFormatter()->setStyle('magenta', new OutputFormatterStyle('magenta', null, array('bold')));
+        $this->output->getFormatter()->setStyle('white', new OutputFormatterStyle('white', null, array('bold')));
     }
 
 }
