@@ -1,23 +1,16 @@
 <?php
 namespace Grav\Console\Cli;
 
-use Grav\Console\ConsoleTrait;
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Formatter\OutputFormatterStyle;
+use Grav\Console\ConsoleCommand;
 use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Yaml\Yaml;
 
 /**
  * Class ComposerCommand
  * @package Grav\Console\Cli
  */
-class ComposerCommand extends Command
+class ComposerCommand extends ConsoleCommand
 {
-    use ConsoleTrait;
-
     /**
      * @var
      */
@@ -59,24 +52,19 @@ class ComposerCommand extends Command
     }
 
     /**
-     * @param InputInterface  $input
-     * @param OutputInterface $output
-     *
      * @return int|null|void
      */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function serve()
     {
-        $this->setupConsole($input, $output);
+        $action = $this->input->getOption('install') ? 'install' : ($this->input->getOption('update') ? 'update' : 'install');
 
-        $action = $input->getOption('install') ? 'install' : ($input->getOption('update') ? 'update' : 'install');
-
-        if ($input->getOption('install')) {
+        if ($this->input->getOption('install')) {
             $action = 'install';
         }
 
         // Updates composer first
-        $output->writeln("\nInstalling vendor dependencies");
-        $output->writeln($this->composerUpdate(GRAV_ROOT, $action));
+        $this->output->writeln("\nInstalling vendor dependencies");
+        $this->output->writeln($this->composerUpdate(GRAV_ROOT, $action));
     }
 
 }
