@@ -1,22 +1,17 @@
 <?php
 namespace Grav\Console\Cli;
 
-use Grav\Console\ConsoleTrait;
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Formatter\OutputFormatterStyle;
+use Grav\Console\ConsoleCommand;
 use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Yaml\Yaml;
 
 /**
  * Class InstallCommand
  * @package Grav\Console\Cli
  */
-class InstallCommand extends Command
+class InstallCommand extends ConsoleCommand
 {
-    use ConsoleTrait;
     /**
      * @var
      */
@@ -51,24 +46,18 @@ class InstallCommand extends Command
                 'destination',
                 InputArgument::OPTIONAL,
                 'Where to install the required bits (default to current project)'
-
             )
             ->setDescription("Installs the dependencies needed by Grav. Optionally can create symbolic links")
             ->setHelp('The <info>install</info> command installs the dependencies needed by Grav. Optionally can create symbolic links');
     }
 
     /**
-     * @param InputInterface  $input
-     * @param OutputInterface $output
-     *
      * @return int|null|void
      */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function serve()
     {
-        $this->setupConsole($input, $output);
-
         $dependencies_file = '.dependencies';
-        $this->destination = ($input->getArgument('destination')) ? $input->getArgument('destination') : ROOT_DIR;
+        $this->destination = ($this->input->getArgument('destination')) ? $this->input->getArgument('destination') : ROOT_DIR;
 
         // fix trailing slash
         $this->destination = rtrim($this->destination, DS) . DS;
@@ -78,7 +67,7 @@ class InstallCommand extends Command
             $local_config_file = exec('eval echo ~/.grav/config');
             if (file_exists($local_config_file)) {
                 $this->local_config = Yaml::parse($local_config_file);
-                $output->writeln('Read local config from <cyan>' . $local_config_file . '</cyan>');
+                $this->output->writeln('Read local config from <cyan>' . $local_config_file . '</cyan>');
             }
         }
 
@@ -88,22 +77,22 @@ class InstallCommand extends Command
         } elseif (file_exists($this->destination . $dependencies_file)) {
             $this->config = Yaml::parse($this->destination . $dependencies_file);
         } else {
-            $output->writeln('<red>ERROR</red> Missing .dependencies file in <cyan>user/</cyan> folder');
+            $this->output->writeln('<red>ERROR</red> Missing .dependencies file in <cyan>user/</cyan> folder');
         }
 
         // If yaml config, process
         if ($this->config) {
-            if (!$input->getOption('symlink')) {
+            if (!$this->input->getOption('symlink')) {
                 // Updates composer first
-                $output->writeln("\nInstalling vendor dependencies");
-                $output->writeln($this->composerUpdate(GRAV_ROOT, 'install'));
+                $this->output->writeln("\nInstalling vendor dependencies");
+                $this->output->writeln($this->composerUpdate(GRAV_ROOT, 'install'));
 
                 $this->gitclone();
             } else {
                 $this->symlink();
             }
         } else {
-            $output->writeln('<red>ERROR</red> invalid YAML in ' . $dependencies_file);
+            $this->output->writeln('<red>ERROR</red> invalid YAML in ' . $dependencies_file);
         }
 
 
