@@ -1,23 +1,17 @@
 <?php
 namespace Grav\Console\Cli;
 
+use Grav\Console\ConsoleCommand;
 use Grav\Common\Filesystem\Folder;
-use Grav\Console\ConsoleTrait;
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Formatter\OutputFormatterStyle;
 use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
 
 /**
  * Class SandboxCommand
  * @package Grav\Console\Cli
  */
-class SandboxCommand extends Command
+class SandboxCommand extends ConsoleCommand
 {
-    use ConsoleTrait;
-
     /**
      * @var array
      */
@@ -41,8 +35,6 @@ class SandboxCommand extends Command
     protected $files = array(
         '/.dependencies',
         '/.htaccess',
-        '/nginx.conf',
-        '/web.config',
         '/user/config/site.yaml',
         '/user/config/system.yaml',
     );
@@ -51,16 +43,19 @@ class SandboxCommand extends Command
      * @var array
      */
     protected $mappings = array(
-        '/.editorconfig' => '/.editorconfig',
-        '/.gitignore' => '/.gitignore',
-        '/CHANGELOG.md' => '/CHANGELOG.md',
-        '/LICENSE' => '/LICENSE',
-        '/README.md' => '/README.md',
-        '/index.php'     => '/index.php',
-        '/composer.json' => '/composer.json',
-        '/bin'           => '/bin',
-        '/system'        => '/system',
-        '/vendor'        => '/vendor',
+        '/.editorconfig'        => '/.editorconfig',
+        '/.gitignore'           => '/.gitignore',
+        '/CHANGELOG.md'         => '/CHANGELOG.md',
+        '/LICENSE.txt'          => '/LICENSE.txt',
+        '/README.md'            => '/README.md',
+        '/CONTRIBUTING.md'      => '/CONTRIBUTING.md',
+        '/index.php'            => '/index.php',
+        '/composer.json'        => '/composer.json',
+        '/bin'                  => '/bin',
+        '/system'               => '/system',
+        '/vendor'               => '/vendor',
+        '/webserver-configs'    => '/webserver-configs',
+        '/codeception.yml'      => '/codeception.yml',
     );
 
     /**
@@ -96,18 +91,14 @@ class SandboxCommand extends Command
     }
 
     /**
-     * @param InputInterface  $input
-     * @param OutputInterface $output
-     *
      * @return int|null|void
      */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function serve()
     {
-        $this->setupConsole($input, $output);
-        $this->destination = $input->getArgument('destination');
+        $this->destination = $this->input->getArgument('destination');
 
         // Symlink the Core Stuff
-        if ($input->getOption('symlink')) {
+        if ($this->input->getOption('symlink')) {
             // Create Some core stuff if it doesn't exist
             $this->createDirectories();
 
@@ -172,7 +163,7 @@ class SandboxCommand extends Command
             $to = $this->destination . $target;
 
             $this->output->writeln('    <cyan>' . $source . '</cyan> <comment>-></comment> ' . $to);
-            Folder::rcopy($from, $to);
+            @Folder::rcopy($from, $to);
         }
     }
 
