@@ -2,21 +2,17 @@
 namespace Grav\Console\Gpm;
 
 use Grav\Common\GPM\GPM;
-use Grav\Console\ConsoleTrait;
-use Symfony\Component\Console\Command\Command;
+use Grav\Console\ConsoleCommand;
 use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 /**
  * Class InfoCommand
  * @package Grav\Console\Gpm
  */
-class InfoCommand extends Command
+class InfoCommand extends ConsoleCommand
 {
-    use ConsoleTrait;
-
     /**
      * @var
      */
@@ -39,6 +35,12 @@ class InfoCommand extends Command
                 InputOption::VALUE_NONE,
                 'Force fetching the new data remotely'
             )
+            ->addOption(
+                'all-yes',
+                'y',
+                InputOption::VALUE_NONE,
+                'Assumes yes (or best approach) instead of prompting'
+            )
             ->addArgument(
                 'package',
                 InputArgument::REQUIRED,
@@ -49,20 +51,16 @@ class InfoCommand extends Command
     }
 
     /**
-     * @param InputInterface  $input
-     * @param OutputInterface $output
-     *
      * @return int|null|void
      */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function serve()
     {
-        $this->setupConsole($input, $output);
         $this->gpm = new GPM($this->input->getOption('force'));
 
-        $foundPackage = $this->gpm->findPackage($input->getArgument('package'));
+        $foundPackage = $this->gpm->findPackage($this->input->getArgument('package'));
 
         if (!$foundPackage) {
-            $this->output->writeln("The package <cyan>'" . $input->getArgument('package') . "'</cyan> was not found in the Grav repository.");
+            $this->output->writeln("The package <cyan>'" . $this->input->getArgument('package') . "'</cyan> was not found in the Grav repository.");
             $this->output->writeln('');
             $this->output->writeln("You can list all the available packages by typing:");
             $this->output->writeln("    <green>" . $this->argv . " index</green>");
@@ -70,7 +68,7 @@ class InfoCommand extends Command
             exit;
         }
 
-        $this->output->writeln("Found package <cyan>'" . $input->getArgument('package') . "'</cyan> under the '<green>" . ucfirst($foundPackage->package_type) . "</green>' section");
+        $this->output->writeln("Found package <cyan>'" . $this->input->getArgument('package') . "'</cyan> under the '<green>" . ucfirst($foundPackage->package_type) . "</green>' section");
         $this->output->writeln('');
         $this->output->writeln("<cyan>" . $foundPackage->name . "</cyan> [" . $foundPackage->slug . "]");
         $this->output->writeln(str_repeat('-', strlen($foundPackage->name) + strlen($foundPackage->slug) + 3));
@@ -116,9 +114,62 @@ class InfoCommand extends Command
             }
         }
 
+        $type = rtrim($foundPackage->package_type, 's');
+        $updatable = $this->gpm->{'is' . $type . 'Updatable'}($foundPackage->slug);
+        $installed = $this->gpm->{'is' . $type . 'Installed'}($foundPackage->slug);
+
+        // display current version if installed and different
+        if ($installed && $updatable) {
+            $local = $this->gpm->{'getInstalled'. $type}($foundPackage->slug);
+            $this->output->writeln('');
+            $this->output->writeln("Currently installed version: <magenta>" . $local->version . "</magenta>");
+            $this->output->writeln('');
+        }
+
+        // display changelog information
+        $questionHelper = $this->getHelper('question');
+        $skipPrompt = $this->input->getOption('all-yes');
+
+        if (!$skipPrompt) {
+            $question = new ConfirmationQuestion("Would you like to read the changelog? [y|N] ",
+                false);
+            $answer = $questionHelper->ask($this->input, $this->output, $question);
+
+            if ($answer) {
+                $changelog = $foundPackage->changelog;
+
+                $this->output->writeln("");
+                foreach ($changelog as $version => $log) {
+                    $title = $version . ' [' . $log['date'] . ']';
+                    $content = preg_replace_callback("/\d\.\s\[\]\(#(.*)\)/", function ($match) {
+                        return "\n" . ucfirst($match[1]) . ":";
+                    }, $log['content']);
+
+                    $this->output->writeln('<cyan>'.$title.'</cyan>');
+                    $this->output->writeln(str_repeat('-', strlen($title)));
+                    $this->output->writeln($content);
+                    $this->output->writeln("");
+
+                    $question = new ConfirmationQuestion("Press [ENTER] to continue or [q] to quit ", true);
+                    if (!$questionHelper->ask($this->input, $this->output, $question)) {
+                        break;
+                    }
+                    $this->output->writeln("");
+                }
+            }
+        }
+
+
         $this->output->writeln('');
-        $this->output->writeln("You can install this package by typing:");
-        $this->output->writeln("    <green>" . $this->argv . " install</green> <cyan>" . $foundPackage->slug . "</cyan>");
+
+        if ($installed && $updatable) {
+            $this->output->writeln("You can update this package by typing:");
+            $this->output->writeln("    <green>" . $this->argv . " update</green> <cyan>" . $foundPackage->slug . "</cyan>");
+        } else {
+            $this->output->writeln("You can install this package by typing:");
+            $this->output->writeln("    <green>" . $this->argv . " install</green> <cyan>" . $foundPackage->slug . "</cyan>");
+        }
+
         $this->output->writeln('');
 
     }
